@@ -51,25 +51,34 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3"
+        "php": ">=5.6.1",
+        "paragonie/constant_time_encoding": "^1|^2",
+        "paragonie/random_compat": "^1.4|^2.0|^9.99.99"
     },
     "require-dev": {
-        "phing/phing": "~2.7",
-        "phpunit/phpunit": "^4.8.35|^5.7|^6.0",
-        "squizlabs/php_codesniffer": "~2.0"
+        "phpunit/phpunit": "*"
     },
     "suggest": {
         "ext-libsodium": "SSH2/SFTP can make use of some algorithms provided by the libsodium-php extension.",
         "ext-openssl": "Install the OpenSSL extension in order to speed up a wide variety of cryptographic operations.",
         "ext-mcrypt": "Install the Mcrypt extension in order to speed up a few other cryptographic operations.",
-        "ext-gmp": "Install the GMP (GNU Multiple Precision) extension in order to speed up arbitrary precision integer arithmetic operations."
+        "ext-gmp": "Install the GMP (GNU Multiple Precision) extension in order to speed up arbitrary precision integer arithmetic operations.",
+        "ext-dom": "Install the DOM extension to load XML formatted public keys."
     },
     "autoload": {
         "files": [
             "phpseclib/bootstrap.php"
         ],
         "psr-4": {
-            "phpseclib\\": "phpseclib/"
+            "phpseclib3\\": "phpseclib/"
         }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "phpseclib3\\Tests\\": "tests/"
+        }
+    },
+    "config": {
+        "sort-packages": true
     }
 }
