@@ -18,7 +18,7 @@
  * <?php
  *    include 'vendor/autoload.php';
  *
- *    $des = new \phpseclib\Crypt\DES();
+ *    $des = new \phpseclib3\Crypt\DES('ctr');
  *
  *    $des->setKey('abcdefgh');
  *
@@ -32,120 +32,119 @@
  * ?>
  * </code>
  *
- * @category  Crypt
- * @package   DES
  * @author    Jim Wigginton <terrafrost@php.net>
  * @copyright 2007 Jim Wigginton
  * @license   http://www.opensource.org/licenses/mit-license.html  MIT License
  * @link      http://phpseclib.sourceforge.net
  */
 
-namespace phpseclib\Crypt;
+namespace phpseclib3\Crypt;
+
+use phpseclib3\Crypt\Common\BlockCipher;
+use phpseclib3\Exception\BadModeException;
 
 /**
  * Pure-PHP implementation of DES.
  *
- * @package DES
  * @author  Jim Wigginton <terrafrost@php.net>
- * @access  public
  */
-class DES extends Base
+class DES extends BlockCipher
 {
-    /**#@+
-     * @access private
-     * @see \phpseclib\Crypt\DES::_setupKey()
-     * @see \phpseclib\Crypt\DES::_processBlock()
-     */
     /**
      * Contains $keys[self::ENCRYPT]
+     *
+     * @see \phpseclib3\Crypt\DES::setupKey()
+     * @see \phpseclib3\Crypt\DES::processBlock()
      */
     const ENCRYPT = 0;
     /**
      * Contains $keys[self::DECRYPT]
+     *
+     * @see \phpseclib3\Crypt\DES::setupKey()
+     * @see \phpseclib3\Crypt\DES::processBlock()
      */
     const DECRYPT = 1;
-    /**#@-*/
 
     /**
      * Block Length of the cipher
      *
-     * @see \phpseclib\Crypt\Base::block_size
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::block_size
      * @var int
-     * @access private
      */
-    var $block_size = 8;
+    protected $block_size = 8;
 
     /**
      * Key Length (in bytes)
      *
-     * @see \phpseclib\Crypt\Base::setKeyLength()
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::setKeyLength()
      * @var int
-     * @access private
      */
-    var $key_length = 8;
+    protected $key_length = 8;
 
     /**
      * The mcrypt specific name of the cipher
      *
-     * @see \phpseclib\Crypt\Base::cipher_name_mcrypt
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::cipher_name_mcrypt
      * @var string
-     * @access private
      */
-    var $cipher_name_mcrypt = 'des';
+    protected $cipher_name_mcrypt = 'des';
 
     /**
      * The OpenSSL names of the cipher / modes
      *
-     * @see \phpseclib\Crypt\Base::openssl_mode_names
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::openssl_mode_names
      * @var array
-     * @access private
      */
-    var $openssl_mode_names = array(
+    protected $openssl_mode_names = [
         self::MODE_ECB => 'des-ecb',
         self::MODE_CBC => 'des-cbc',
         self::MODE_CFB => 'des-cfb',
         self::MODE_OFB => 'des-ofb'
         // self::MODE_CTR is undefined for DES
-    );
+    ];
 
     /**
      * Optimizing value while CFB-encrypting
      *
-     * @see \phpseclib\Crypt\Base::cfb_init_len
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::cfb_init_len
      * @var int
-     * @access private
      */
-    var $cfb_init_len = 500;
+    protected $cfb_init_len = 500;
 
     /**
      * Switch for DES/3DES encryption
      *
      * Used only if $engine == self::ENGINE_INTERNAL
      *
-     * @see self::_setupKey()
-     * @see self::_processBlock()
+     * @see self::setupKey()
+     * @see self::processBlock()
      * @var int
-     * @access private
      */
-    var $des_rounds = 1;
+    protected $des_rounds = 1;
 
     /**
      * max possible size of $key
      *
      * @see self::setKey()
      * @var string
-     * @access private
      */
-    var $key_length_max = 8;
+    protected $key_length_max = 8;
 
     /**
      * The Key Schedule
      *
-     * @see self::_setupKey()
+     * @see self::setupKey()
      * @var array
-     * @access private
      */
-    var $keys;
+    private $keys;
+
+    /**
+     * Key Cache "key"
+     *
+     * @see self::setupKey()
+     * @var array
+     */
+    private $kl;
 
     /**
      * Shuffle table.
@@ -154,12 +153,11 @@ class DES extends Base
      * with each byte containing all bits in the same state as the
      * corresponding bit in the index value.
      *
-     * @see self::_processBlock()
-     * @see self::_setupKey()
+     * @see self::processBlock()
+     * @see self::setupKey()
      * @var array
-     * @access private
      */
-    var $shuffle = array(
+    protected static $shuffle = [
         "\x00\x00\x00\x00\x00\x00\x00\x00", "\x00\x00\x00\x00\x00\x00\x00\xFF",
         "\x00\x00\x00\x00\x00\x00\xFF\x00", "\x00\x00\x00\x00\x00\x00\xFF\xFF",
         "\x00\x00\x00\x00\x00\xFF\x00\x00", "\x00\x00\x00\x00\x00\xFF\x00\xFF",
@@ -288,7 +286,7 @@ class DES extends Base
         "\xFF\xFF\xFF\xFF\xFF\x00\xFF\x00", "\xFF\xFF\xFF\xFF\xFF\x00\xFF\xFF",
         "\xFF\xFF\xFF\xFF\xFF\xFF\x00\x00", "\xFF\xFF\xFF\xFF\xFF\xFF\x00\xFF",
         "\xFF\xFF\xFF\xFF\xFF\xFF\xFF\x00", "\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF"
-    );
+    ];
 
     /**
      * IP mapping helper table.
@@ -296,9 +294,8 @@ class DES extends Base
      * Indexing this table with each source byte performs the initial bit permutation.
      *
      * @var array
-     * @access private
      */
-    var $ipmap = array(
+    protected static $ipmap = [
         0x00, 0x10, 0x01, 0x11, 0x20, 0x30, 0x21, 0x31,
         0x02, 0x12, 0x03, 0x13, 0x22, 0x32, 0x23, 0x33,
         0x40, 0x50, 0x41, 0x51, 0x60, 0x70, 0x61, 0x71,
@@ -331,16 +328,15 @@ class DES extends Base
         0x8E, 0x9E, 0x8F, 0x9F, 0xAE, 0xBE, 0xAF, 0xBF,
         0xCC, 0xDC, 0xCD, 0xDD, 0xEC, 0xFC, 0xED, 0xFD,
         0xCE, 0xDE, 0xCF, 0xDF, 0xEE, 0xFE, 0xEF, 0xFF
-    );
+    ];
 
     /**
      * Inverse IP mapping helper table.
      * Indexing this table with a byte value reverses the bit order.
      *
      * @var array
-     * @access private
      */
-    var $invipmap = array(
+    protected static $invipmap = [
         0x00, 0x80, 0x40, 0xC0, 0x20, 0xA0, 0x60, 0xE0,
         0x10, 0x90, 0x50, 0xD0, 0x30, 0xB0, 0x70, 0xF0,
         0x08, 0x88, 0x48, 0xC8, 0x28, 0xA8, 0x68, 0xE8,
@@ -373,7 +369,7 @@ class DES extends Base
         0x17, 0x97, 0x57, 0xD7, 0x37, 0xB7, 0x77, 0xF7,
         0x0F, 0x8F, 0x4F, 0xCF, 0x2F, 0xAF, 0x6F, 0xEF,
         0x1F, 0x9F, 0x5F, 0xDF, 0x3F, 0xBF, 0x7F, 0xFF
-    );
+    ];
 
     /**
      * Pre-permuted S-box1
@@ -382,9 +378,8 @@ class DES extends Base
      * P table: concatenation can then be replaced by exclusive ORs.
      *
      * @var array
-     * @access private
      */
-    var $sbox1 = array(
+    protected static $sbox1 = [
         0x00808200, 0x00000000, 0x00008000, 0x00808202,
         0x00808002, 0x00008202, 0x00000002, 0x00008000,
         0x00000200, 0x00808200, 0x00808202, 0x00000200,
@@ -401,15 +396,14 @@ class DES extends Base
         0x00800002, 0x00000202, 0x00008202, 0x00808200,
         0x00000202, 0x00800200, 0x00800200, 0x00000000,
         0x00008002, 0x00008200, 0x00000000, 0x00808002
-    );
+    ];
 
     /**
      * Pre-permuted S-box2
      *
      * @var array
-     * @access private
      */
-    var $sbox2 = array(
+    protected static $sbox2 = [
         0x40084010, 0x40004000, 0x00004000, 0x00084010,
         0x00080000, 0x00000010, 0x40080010, 0x40004010,
         0x40000010, 0x40084010, 0x40084000, 0x40000000,
@@ -426,15 +420,14 @@ class DES extends Base
         0x00080010, 0x40004010, 0x40000010, 0x00080010,
         0x00084000, 0x00000000, 0x40004000, 0x00004010,
         0x40000000, 0x40080010, 0x40084010, 0x00084000
-    );
+    ];
 
     /**
      * Pre-permuted S-box3
      *
      * @var array
-     * @access private
      */
-    var $sbox3 = array(
+    protected static $sbox3 = [
         0x00000104, 0x04010100, 0x00000000, 0x04010004,
         0x04000100, 0x00000000, 0x00010104, 0x04000100,
         0x00010004, 0x04000004, 0x04000004, 0x00010000,
@@ -451,15 +444,14 @@ class DES extends Base
         0x00000004, 0x00010104, 0x00010100, 0x04000004,
         0x04010000, 0x04000104, 0x00000104, 0x04010000,
         0x00010104, 0x00000004, 0x04010004, 0x00010100
-    );
+    ];
 
     /**
      * Pre-permuted S-box4
      *
      * @var array
-     * @access private
      */
-    var $sbox4 = array(
+    protected static $sbox4 = [
         0x80401000, 0x80001040, 0x80001040, 0x00000040,
         0x00401040, 0x80400040, 0x80400000, 0x80001000,
         0x00000000, 0x00401000, 0x00401000, 0x80401040,
@@ -476,15 +468,14 @@ class DES extends Base
         0x80400000, 0x80001000, 0x00401040, 0x80400040,
         0x80001000, 0x00001040, 0x00400000, 0x80401000,
         0x00000040, 0x00400000, 0x00001000, 0x00401040
-    );
+    ];
 
     /**
      * Pre-permuted S-box5
      *
      * @var array
-     * @access private
      */
-    var $sbox5 = array(
+    protected static $sbox5 = [
         0x00000080, 0x01040080, 0x01040000, 0x21000080,
         0x00040000, 0x00000080, 0x20000000, 0x01040000,
         0x20040080, 0x00040000, 0x01000080, 0x20040080,
@@ -501,15 +492,14 @@ class DES extends Base
         0x01040000, 0x00000000, 0x20040000, 0x21000000,
         0x00040080, 0x01000080, 0x20000080, 0x00040000,
         0x00000000, 0x20040000, 0x01040080, 0x20000080
-    );
+    ];
 
     /**
      * Pre-permuted S-box6
      *
      * @var array
-     * @access private
      */
-    var $sbox6 = array(
+    protected static $sbox6 = [
         0x10000008, 0x10200000, 0x00002000, 0x10202008,
         0x10200000, 0x00000008, 0x10202008, 0x00200000,
         0x10002000, 0x00202008, 0x00200000, 0x10000008,
@@ -526,15 +516,14 @@ class DES extends Base
         0x00000008, 0x00002000, 0x10200000, 0x00202008,
         0x00002000, 0x00200008, 0x10002008, 0x00000000,
         0x10202000, 0x10000000, 0x00200008, 0x10002008
-    );
+    ];
 
     /**
      * Pre-permuted S-box7
      *
      * @var array
-     * @access private
      */
-    var $sbox7 = array(
+    protected static $sbox7 = [
         0x00100000, 0x02100001, 0x02000401, 0x00000000,
         0x00000400, 0x02000401, 0x00100401, 0x02100400,
         0x02100401, 0x00100000, 0x00000000, 0x02000001,
@@ -551,15 +540,14 @@ class DES extends Base
         0x00100400, 0x00000000, 0x00000001, 0x02100401,
         0x00000000, 0x00100401, 0x02100000, 0x00000400,
         0x02000001, 0x02000400, 0x00000400, 0x00100001
-    );
+    ];
 
     /**
      * Pre-permuted S-box8
      *
      * @var array
-     * @access private
      */
-    var $sbox8 = array(
+    protected static $sbox8 = [
         0x08000820, 0x00000800, 0x00020000, 0x08020820,
         0x08000000, 0x08000820, 0x00000020, 0x08000000,
         0x00020020, 0x08020000, 0x08020820, 0x00020800,
@@ -576,51 +564,64 @@ class DES extends Base
         0x08020000, 0x08000800, 0x08000820, 0x00000000,
         0x08020820, 0x00020800, 0x00020800, 0x00000820,
         0x00000820, 0x00020020, 0x08000000, 0x08020800
-    );
+    ];
+
+    /**
+     * Default Constructor.
+     *
+     * @param string $mode
+     * @throws BadModeException if an invalid / unsupported mode is provided
+     */
+    public function __construct($mode)
+    {
+        parent::__construct($mode);
+
+        if ($this->mode == self::MODE_STREAM) {
+            throw new BadModeException('Block ciphers cannot be ran in stream mode');
+        }
+    }
 
     /**
      * Test for engine validity
      *
-     * This is mainly just a wrapper to set things up for \phpseclib\Crypt\Base::isValidEngine()
+     * This is mainly just a wrapper to set things up for \phpseclib3\Crypt\Common\SymmetricKey::isValidEngine()
      *
-     * @see \phpseclib\Crypt\Base::isValidEngine()
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::isValidEngine()
      * @param int $engine
-     * @access public
      * @return bool
      */
-    function isValidEngine($engine)
+    protected function isValidEngineHelper($engine)
     {
         if ($this->key_length_max == 8) {
             if ($engine == self::ENGINE_OPENSSL) {
+                // quoting https://www.openssl.org/news/openssl-3.0-notes.html, OpenSSL 3.0.1
+                // "Moved all variations of the EVP ciphers CAST5, BF, IDEA, SEED, RC2, RC4, RC5, and DES to the legacy provider"
+                // in theory openssl_get_cipher_methods() should catch this but, on GitHub Actions, at least, it does not
+                if (defined('OPENSSL_VERSION_TEXT') && version_compare(preg_replace('#OpenSSL (\d+\.\d+\.\d+) .*#', '$1', OPENSSL_VERSION_TEXT), '3.0.1', '>=')) {
+                    return false;
+                }
                 $this->cipher_name_openssl_ecb = 'des-ecb';
-                $this->cipher_name_openssl = 'des-' . $this->_openssl_translate_mode();
+                $this->cipher_name_openssl = 'des-' . $this->openssl_translate_mode();
             }
         }
 
-        return parent::isValidEngine($engine);
+        return parent::isValidEngineHelper($engine);
     }
 
     /**
      * Sets the key.
      *
-     * Keys can be of any length.  DES, itself, uses 64-bit keys (eg. strlen($key) == 8), however, we
-     * only use the first eight, if $key has more then eight characters in it, and pad $key with the
-     * null byte if it is less then eight characters long.
+     * Keys must be 64-bits long or 8 bytes long.
      *
      * DES also requires that every eighth bit be a parity bit, however, we'll ignore that.
      *
-     * If the key is not explicitly set, it'll be assumed to be all zero's.
-     *
-     * @see \phpseclib\Crypt\Base::setKey()
-     * @access public
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::setKey()
      * @param string $key
      */
-    function setKey($key)
+    public function setKey($key)
     {
-        // We check/cut here only up to max length of the key.
-        // Key padding to the proper length will be done in _setupKey()
-        if (strlen($key) > $this->key_length_max) {
-            $key = substr($key, 0, $this->key_length_max);
+        if (!($this instanceof TripleDES) && strlen($key) != 8) {
+            throw new \LengthException('Key of size ' . strlen($key) . ' not supported by this algorithm. Only keys of size 8 are supported');
         }
 
         // Sets the key
@@ -630,31 +631,29 @@ class DES extends Base
     /**
      * Encrypts a block
      *
-     * @see \phpseclib\Crypt\Base::_encryptBlock()
-     * @see \phpseclib\Crypt\Base::encrypt()
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::encryptBlock()
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::encrypt()
      * @see self::encrypt()
-     * @access private
      * @param string $in
      * @return string
      */
-    function _encryptBlock($in)
+    protected function encryptBlock($in)
     {
-        return $this->_processBlock($in, self::ENCRYPT);
+        return $this->processBlock($in, self::ENCRYPT);
     }
 
     /**
      * Decrypts a block
      *
-     * @see \phpseclib\Crypt\Base::_decryptBlock()
-     * @see \phpseclib\Crypt\Base::decrypt()
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::decryptBlock()
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::decrypt()
      * @see self::decrypt()
-     * @access private
      * @param string $in
      * @return string
      */
-    function _decryptBlock($in)
+    protected function decryptBlock($in)
     {
-        return $this->_processBlock($in, self::DECRYPT);
+        return $this->processBlock($in, self::DECRYPT);
     }
 
     /**
@@ -664,29 +663,28 @@ class DES extends Base
      * {@link http://en.wikipedia.org/wiki/Image:Feistel.png Feistel.png} to get a general
      * idea of what this function does.
      *
-     * @see self::_encryptBlock()
-     * @see self::_decryptBlock()
-     * @access private
+     * @see self::encryptBlock()
+     * @see self::decryptBlock()
      * @param string $block
      * @param int $mode
      * @return string
      */
-    function _processBlock($block, $mode)
+    private function processBlock($block, $mode)
     {
         static $sbox1, $sbox2, $sbox3, $sbox4, $sbox5, $sbox6, $sbox7, $sbox8, $shuffleip, $shuffleinvip;
         if (!$sbox1) {
-            $sbox1 = array_map("intval", $this->sbox1);
-            $sbox2 = array_map("intval", $this->sbox2);
-            $sbox3 = array_map("intval", $this->sbox3);
-            $sbox4 = array_map("intval", $this->sbox4);
-            $sbox5 = array_map("intval", $this->sbox5);
-            $sbox6 = array_map("intval", $this->sbox6);
-            $sbox7 = array_map("intval", $this->sbox7);
-            $sbox8 = array_map("intval", $this->sbox8);
+            $sbox1 = array_map('intval', self::$sbox1);
+            $sbox2 = array_map('intval', self::$sbox2);
+            $sbox3 = array_map('intval', self::$sbox3);
+            $sbox4 = array_map('intval', self::$sbox4);
+            $sbox5 = array_map('intval', self::$sbox5);
+            $sbox6 = array_map('intval', self::$sbox6);
+            $sbox7 = array_map('intval', self::$sbox7);
+            $sbox8 = array_map('intval', self::$sbox8);
             /* Merge $shuffle with $[inv]ipmap */
             for ($i = 0; $i < 256; ++$i) {
-                $shuffleip[]    =  $this->shuffle[$this->ipmap[$i]];
-                $shuffleinvip[] =  $this->shuffle[$this->invipmap[$i]];
+                $shuffleip[]    =  self::$shuffle[self::$ipmap[$i]];
+                $shuffleinvip[] =  self::$shuffle[self::$invipmap[$i]];
             }
         }
 
@@ -695,7 +693,7 @@ class DES extends Base
 
         // Do the initial IP permutation.
         $t = unpack('Nl/Nr', $block);
-        list($l, $r) = array($t['l'], $t['r']);
+        list($l, $r) = [$t['l'], $t['r']];
         $block = ($shuffleip[ $r        & 0xFF] & "\x80\x80\x80\x80\x80\x80\x80\x80") |
                  ($shuffleip[($r >>  8) & 0xFF] & "\x40\x40\x40\x40\x40\x40\x40\x40") |
                  ($shuffleip[($r >> 16) & 0xFF] & "\x20\x20\x20\x20\x20\x20\x20\x20") |
@@ -707,7 +705,7 @@ class DES extends Base
 
         // Extract L0 and R0.
         $t = unpack('Nl/Nr', $block);
-        list($l, $r) = array($t['l'], $t['r']);
+        list($l, $r) = [$t['l'], $t['r']];
 
         for ($des_round = 0; $des_round < $this->des_rounds; ++$des_round) {
             // Perform the 16 steps.
@@ -749,22 +747,21 @@ class DES extends Base
     /**
      * Creates the key schedule
      *
-     * @see \phpseclib\Crypt\Base::_setupKey()
-     * @access private
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::setupKey()
      */
-    function _setupKey()
+    protected function setupKey()
     {
         if (isset($this->kl['key']) && $this->key === $this->kl['key'] && $this->des_rounds === $this->kl['des_rounds']) {
             // already expanded
             return;
         }
-        $this->kl = array('key' => $this->key, 'des_rounds' => $this->des_rounds);
+        $this->kl = ['key' => $this->key, 'des_rounds' => $this->des_rounds];
 
-        static $shifts = array( // number of key bits shifted per round
+        static $shifts = [ // number of key bits shifted per round
             1, 1, 2, 2, 2, 2, 2, 2, 1, 2, 2, 2, 2, 2, 2, 1
-        );
+        ];
 
-        static $pc1map = array(
+        static $pc1map = [
             0x00, 0x00, 0x08, 0x08, 0x04, 0x04, 0x0C, 0x0C,
             0x02, 0x02, 0x0A, 0x0A, 0x06, 0x06, 0x0E, 0x0E,
             0x10, 0x10, 0x18, 0x18, 0x14, 0x14, 0x1C, 0x1C,
@@ -797,16 +794,16 @@ class DES extends Base
             0xE2, 0xE2, 0xEA, 0xEA, 0xE6, 0xE6, 0xEE, 0xEE,
             0xF0, 0xF0, 0xF8, 0xF8, 0xF4, 0xF4, 0xFC, 0xFC,
             0xF2, 0xF2, 0xFA, 0xFA, 0xF6, 0xF6, 0xFE, 0xFE
-        );
+        ];
 
         // Mapping tables for the PC-2 transformation.
-        static $pc2mapc1 = array(
+        static $pc2mapc1 = [
             0x00000000, 0x00000400, 0x00200000, 0x00200400,
             0x00000001, 0x00000401, 0x00200001, 0x00200401,
             0x02000000, 0x02000400, 0x02200000, 0x02200400,
             0x02000001, 0x02000401, 0x02200001, 0x02200401
-        );
-        static $pc2mapc2 = array(
+        ];
+        static $pc2mapc2 = [
             0x00000000, 0x00000800, 0x08000000, 0x08000800,
             0x00010000, 0x00010800, 0x08010000, 0x08010800,
             0x00000000, 0x00000800, 0x08000000, 0x08000800,
@@ -871,8 +868,8 @@ class DES extends Base
             0x01050110, 0x01050910, 0x09050110, 0x09050910,
             0x01040110, 0x01040910, 0x09040110, 0x09040910,
             0x01050110, 0x01050910, 0x09050110, 0x09050910
-        );
-        static $pc2mapc3 = array(
+        ];
+        static $pc2mapc3 = [
             0x00000000, 0x00000004, 0x00001000, 0x00001004,
             0x00000000, 0x00000004, 0x00001000, 0x00001004,
             0x10000000, 0x10000004, 0x10001000, 0x10001004,
@@ -937,8 +934,8 @@ class DES extends Base
             0x20080022, 0x20080026, 0x20081022, 0x20081026,
             0x30080022, 0x30080026, 0x30081022, 0x30081026,
             0x30080022, 0x30080026, 0x30081022, 0x30081026
-        );
-        static $pc2mapc4 = array(
+        ];
+        static $pc2mapc4 = [
             0x00000000, 0x00100000, 0x00000008, 0x00100008,
             0x00000200, 0x00100200, 0x00000208, 0x00100208,
             0x00000000, 0x00100000, 0x00000008, 0x00100008,
@@ -1003,14 +1000,14 @@ class DES extends Base
             0x04022200, 0x04122200, 0x04022208, 0x04122208,
             0x04022000, 0x04122000, 0x04022008, 0x04122008,
             0x04022200, 0x04122200, 0x04022208, 0x04122208
-        );
-        static $pc2mapd1 = array(
+        ];
+        static $pc2mapd1 = [
             0x00000000, 0x00000001, 0x08000000, 0x08000001,
             0x00200000, 0x00200001, 0x08200000, 0x08200001,
             0x00000002, 0x00000003, 0x08000002, 0x08000003,
             0x00200002, 0x00200003, 0x08200002, 0x08200003
-        );
-        static $pc2mapd2 = array(
+        ];
+        static $pc2mapd2 = [
             0x00000000, 0x00100000, 0x00000800, 0x00100800,
             0x00000000, 0x00100000, 0x00000800, 0x00100800,
             0x04000000, 0x04100000, 0x04000800, 0x04100800,
@@ -1075,8 +1072,8 @@ class DES extends Base
             0x00020204, 0x00120204, 0x00020A04, 0x00120A04,
             0x04020204, 0x04120204, 0x04020A04, 0x04120A04,
             0x04020204, 0x04120204, 0x04020A04, 0x04120A04
-        );
-        static $pc2mapd3 = array(
+        ];
+        static $pc2mapd3 = [
             0x00000000, 0x00010000, 0x02000000, 0x02010000,
             0x00000020, 0x00010020, 0x02000020, 0x02010020,
             0x00040000, 0x00050000, 0x02040000, 0x02050000,
@@ -1141,8 +1138,8 @@ class DES extends Base
             0x20002030, 0x20012030, 0x22002030, 0x22012030,
             0x20042010, 0x20052010, 0x22042010, 0x22052010,
             0x20042030, 0x20052030, 0x22042030, 0x22052030
-        );
-        static $pc2mapd4 = array(
+        ];
+        static $pc2mapd4 = [
             0x00000000, 0x00000400, 0x01000000, 0x01000400,
             0x00000000, 0x00000400, 0x01000000, 0x01000400,
             0x00000100, 0x00000500, 0x01000100, 0x01000500,
@@ -1207,33 +1204,33 @@ class DES extends Base
             0x10081008, 0x10081408, 0x11081008, 0x11081408,
             0x10081108, 0x10081508, 0x11081108, 0x11081508,
             0x10081108, 0x10081508, 0x11081108, 0x11081508
-        );
+        ];
 
-        $keys = array();
+        $keys = [];
         for ($des_round = 0; $des_round < $this->des_rounds; ++$des_round) {
             // pad the key and remove extra characters as appropriate.
             $key = str_pad(substr($this->key, $des_round * 8, 8), 8, "\0");
 
             // Perform the PC/1 transformation and compute C and D.
             $t = unpack('Nl/Nr', $key);
-            list($l, $r) = array($t['l'], $t['r']);
-            $key = ($this->shuffle[$pc1map[ $r        & 0xFF]] & "\x80\x80\x80\x80\x80\x80\x80\x00") |
-                   ($this->shuffle[$pc1map[($r >>  8) & 0xFF]] & "\x40\x40\x40\x40\x40\x40\x40\x00") |
-                   ($this->shuffle[$pc1map[($r >> 16) & 0xFF]] & "\x20\x20\x20\x20\x20\x20\x20\x00") |
-                   ($this->shuffle[$pc1map[($r >> 24) & 0xFF]] & "\x10\x10\x10\x10\x10\x10\x10\x00") |
-                   ($this->shuffle[$pc1map[ $l        & 0xFF]] & "\x08\x08\x08\x08\x08\x08\x08\x00") |
-                   ($this->shuffle[$pc1map[($l >>  8) & 0xFF]] & "\x04\x04\x04\x04\x04\x04\x04\x00") |
-                   ($this->shuffle[$pc1map[($l >> 16) & 0xFF]] & "\x02\x02\x02\x02\x02\x02\x02\x00") |
-                   ($this->shuffle[$pc1map[($l >> 24) & 0xFF]] & "\x01\x01\x01\x01\x01\x01\x01\x00");
+            list($l, $r) = [$t['l'], $t['r']];
+            $key = (self::$shuffle[$pc1map[ $r        & 0xFF]] & "\x80\x80\x80\x80\x80\x80\x80\x00") |
+                   (self::$shuffle[$pc1map[($r >>  8) & 0xFF]] & "\x40\x40\x40\x40\x40\x40\x40\x00") |
+                   (self::$shuffle[$pc1map[($r >> 16) & 0xFF]] & "\x20\x20\x20\x20\x20\x20\x20\x00") |
+                   (self::$shuffle[$pc1map[($r >> 24) & 0xFF]] & "\x10\x10\x10\x10\x10\x10\x10\x00") |
+                   (self::$shuffle[$pc1map[ $l        & 0xFF]] & "\x08\x08\x08\x08\x08\x08\x08\x00") |
+                   (self::$shuffle[$pc1map[($l >>  8) & 0xFF]] & "\x04\x04\x04\x04\x04\x04\x04\x00") |
+                   (self::$shuffle[$pc1map[($l >> 16) & 0xFF]] & "\x02\x02\x02\x02\x02\x02\x02\x00") |
+                   (self::$shuffle[$pc1map[($l >> 24) & 0xFF]] & "\x01\x01\x01\x01\x01\x01\x01\x00");
             $key = unpack('Nc/Nd', $key);
             $c = ( $key['c'] >> 4) & 0x0FFFFFFF;
             $d = (($key['d'] >> 4) & 0x0FFFFFF0) | ($key['c'] & 0x0F);
 
-            $keys[$des_round] = array(
-                self::ENCRYPT => array(),
+            $keys[$des_round] = [
+                self::ENCRYPT => [],
                 self::DECRYPT => array_fill(0, 32, 0)
-            );
-            for ($i = 0, $ki = 31; $i < 16; ++$i, $ki-= 2) {
+            ];
+            for ($i = 0, $ki = 31; $i < 16; ++$i, $ki -= 2) {
                 $c <<= $shifts[$i];
                 $c = ($c | ($c >> 28)) & 0x0FFFFFFF;
                 $d <<= $shifts[$i];
@@ -1246,9 +1243,9 @@ class DES extends Base
                       $pc2mapd3[($d >>  8) & 0xFF] | $pc2mapd4[ $d        & 0xFF];
 
                 // Reorder: odd bytes/even bytes. Push the result in key schedule.
-                $val1 = ( $cp        & 0xFF000000) | (($cp <<  8) & 0x00FF0000) |
+                $val1 = ( $cp        & intval(0xFF000000)) | (($cp <<  8) & 0x00FF0000) |
                         (($dp >> 16) & 0x0000FF00) | (($dp >>  8) & 0x000000FF);
-                $val2 = (($cp <<  8) & 0xFF000000) | (($cp << 16) & 0x00FF0000) |
+                $val2 = (($cp <<  8) & intval(0xFF000000)) | (($cp << 16) & 0x00FF0000) |
                         (($dp >>  8) & 0x0000FF00) | ( $dp        & 0x000000FF);
                 $keys[$des_round][self::ENCRYPT][       ] = $val1;
                 $keys[$des_round][self::DECRYPT][$ki - 1] = $val1;
@@ -1259,7 +1256,7 @@ class DES extends Base
 
         switch ($this->des_rounds) {
             case 3: // 3DES keys
-                $this->keys = array(
+                $this->keys = [
                     self::ENCRYPT => array_merge(
                         $keys[0][self::ENCRYPT],
                         $keys[1][self::DECRYPT],
@@ -1270,174 +1267,126 @@ class DES extends Base
                         $keys[1][self::ENCRYPT],
                         $keys[0][self::DECRYPT]
                     )
-                );
+                ];
                 break;
             // case 1: // DES keys
             default:
-                $this->keys = array(
+                $this->keys = [
                     self::ENCRYPT => $keys[0][self::ENCRYPT],
                     self::DECRYPT => $keys[0][self::DECRYPT]
-                );
+                ];
         }
     }
 
     /**
      * Setup the performance-optimized function for de/encrypt()
      *
-     * @see \phpseclib\Crypt\Base::_setupInlineCrypt()
-     * @access private
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::setupInlineCrypt()
      */
-    function _setupInlineCrypt()
+    protected function setupInlineCrypt()
     {
-        $lambda_functions =& self::_getLambdaFunctions();
-
         // Engine configuration for:
         // -  DES ($des_rounds == 1) or
         // - 3DES ($des_rounds == 3)
         $des_rounds = $this->des_rounds;
 
-        // We create max. 10 hi-optimized code for memory reason. Means: For each $key one ultra fast inline-crypt function.
-        // (Currently, for DES, one generated $lambda_function cost on php5.5@32bit ~135kb unfreeable mem and ~230kb on php5.5@64bit)
-        // (Currently, for TripleDES, one generated $lambda_function cost on php5.5@32bit ~240kb unfreeable mem and ~340kb on php5.5@64bit)
-        // After that, we'll still create very fast optimized code but not the hi-ultimative code, for each $mode one
-        $gen_hi_opt_code = (bool)( count($lambda_functions) < 10 );
-
-        // Generation of a unique hash for our generated code
-        $code_hash = "Crypt_DES, $des_rounds, {$this->mode}";
-        if ($gen_hi_opt_code) {
-            // For hi-optimized code, we create for each combination of
-            // $mode, $des_rounds and $this->key its own encrypt/decrypt function.
-            // After max 10 hi-optimized functions, we create generic
-            // (still very fast.. but not ultra) functions for each $mode/$des_rounds
-            // Currently 2 * 5 generic functions will be then max. possible.
-            $code_hash = str_pad($code_hash, 32) . $this->_hashInlineCryptFunction($this->key);
-        }
-
-        // Is there a re-usable $lambda_functions in there? If not, we have to create it.
-        if (!isset($lambda_functions[$code_hash])) {
-            // Init code for both, encrypt and decrypt.
-            $init_crypt = 'static $sbox1, $sbox2, $sbox3, $sbox4, $sbox5, $sbox6, $sbox7, $sbox8, $shuffleip, $shuffleinvip;
-                if (!$sbox1) {
-                    $sbox1 = array_map("intval", $self->sbox1);
-                    $sbox2 = array_map("intval", $self->sbox2);
-                    $sbox3 = array_map("intval", $self->sbox3);
-                    $sbox4 = array_map("intval", $self->sbox4);
-                    $sbox5 = array_map("intval", $self->sbox5);
-                    $sbox6 = array_map("intval", $self->sbox6);
-                    $sbox7 = array_map("intval", $self->sbox7);
-                    $sbox8 = array_map("intval", $self->sbox8);'
-                    /* Merge $shuffle with $[inv]ipmap */ . '
-                    for ($i = 0; $i < 256; ++$i) {
-                        $shuffleip[]    =  $self->shuffle[$self->ipmap[$i]];
-                        $shuffleinvip[] =  $self->shuffle[$self->invipmap[$i]];
-                    }
+        $init_crypt = 'static $sbox1, $sbox2, $sbox3, $sbox4, $sbox5, $sbox6, $sbox7, $sbox8, $shuffleip, $shuffleinvip;
+            if (!$sbox1) {
+                $sbox1 = array_map("intval", self::$sbox1);
+                $sbox2 = array_map("intval", self::$sbox2);
+                $sbox3 = array_map("intval", self::$sbox3);
+                $sbox4 = array_map("intval", self::$sbox4);
+                $sbox5 = array_map("intval", self::$sbox5);
+                $sbox6 = array_map("intval", self::$sbox6);
+                $sbox7 = array_map("intval", self::$sbox7);
+                $sbox8 = array_map("intval", self::$sbox8);'
+                /* Merge $shuffle with $[inv]ipmap */ . '
+                for ($i = 0; $i < 256; ++$i) {
+                    $shuffleip[]    =  self::$shuffle[self::$ipmap[$i]];
+                    $shuffleinvip[] =  self::$shuffle[self::$invipmap[$i]];
                 }
+            }
+        ';
+
+        $k = [
+            self::ENCRYPT => $this->keys[self::ENCRYPT],
+            self::DECRYPT => $this->keys[self::DECRYPT]
+        ];
+        $init_encrypt = '';
+        $init_decrypt = '';
+
+        // Creating code for en- and decryption.
+        $crypt_block = [];
+        foreach ([self::ENCRYPT, self::DECRYPT] as $c) {
+            /* Do the initial IP permutation. */
+            $crypt_block[$c] = '
+                $in = unpack("N*", $in);
+                $l  = $in[1];
+                $r  = $in[2];
+                $in = unpack("N*",
+                    ($shuffleip[ $r        & 0xFF] & "\x80\x80\x80\x80\x80\x80\x80\x80") |
+                    ($shuffleip[($r >>  8) & 0xFF] & "\x40\x40\x40\x40\x40\x40\x40\x40") |
+                    ($shuffleip[($r >> 16) & 0xFF] & "\x20\x20\x20\x20\x20\x20\x20\x20") |
+                    ($shuffleip[($r >> 24) & 0xFF] & "\x10\x10\x10\x10\x10\x10\x10\x10") |
+                    ($shuffleip[ $l        & 0xFF] & "\x08\x08\x08\x08\x08\x08\x08\x08") |
+                    ($shuffleip[($l >>  8) & 0xFF] & "\x04\x04\x04\x04\x04\x04\x04\x04") |
+                    ($shuffleip[($l >> 16) & 0xFF] & "\x02\x02\x02\x02\x02\x02\x02\x02") |
+                    ($shuffleip[($l >> 24) & 0xFF] & "\x01\x01\x01\x01\x01\x01\x01\x01")
+                );
+                ' . /* Extract L0 and R0 */ '
+                $l = $in[1];
+                $r = $in[2];
             ';
 
-            switch (true) {
-                case $gen_hi_opt_code:
-                    // In Hi-optimized code mode, we use our [3]DES key schedule as hardcoded integers.
-                    // No futher initialisation of the $keys schedule is necessary.
-                    // That is the extra performance boost.
-                    $k = array(
-                        self::ENCRYPT => $this->keys[self::ENCRYPT],
-                        self::DECRYPT => $this->keys[self::DECRYPT]
-                    );
-                    $init_encrypt = '';
-                    $init_decrypt = '';
-                    break;
-                default:
-                    // In generic optimized code mode, we have to use, as the best compromise [currently],
-                    // our key schedule as $ke/$kd arrays. (with hardcoded indexes...)
-                    $k = array(
-                        self::ENCRYPT => array(),
-                        self::DECRYPT => array()
-                    );
-                    for ($i = 0, $c = count($this->keys[self::ENCRYPT]); $i < $c; ++$i) {
-                        $k[self::ENCRYPT][$i] = '$ke[' . $i . ']';
-                        $k[self::DECRYPT][$i] = '$kd[' . $i . ']';
-                    }
-                    $init_encrypt = '$ke = $self->keys[$self::ENCRYPT];';
-                    $init_decrypt = '$kd = $self->keys[$self::DECRYPT];';
-                    break;
-            }
+            $l = '$l';
+            $r = '$r';
 
-            // Creating code for en- and decryption.
-            $crypt_block = array();
-            foreach (array(self::ENCRYPT, self::DECRYPT) as $c) {
-                /* Do the initial IP permutation. */
-                $crypt_block[$c] = '
-                    $in = unpack("N*", $in);
-                    $l  = $in[1];
-                    $r  = $in[2];
-                    $in = unpack("N*",
-                        ($shuffleip[ $r        & 0xFF] & "\x80\x80\x80\x80\x80\x80\x80\x80") |
-                        ($shuffleip[($r >>  8) & 0xFF] & "\x40\x40\x40\x40\x40\x40\x40\x40") |
-                        ($shuffleip[($r >> 16) & 0xFF] & "\x20\x20\x20\x20\x20\x20\x20\x20") |
-                        ($shuffleip[($r >> 24) & 0xFF] & "\x10\x10\x10\x10\x10\x10\x10\x10") |
-                        ($shuffleip[ $l        & 0xFF] & "\x08\x08\x08\x08\x08\x08\x08\x08") |
-                        ($shuffleip[($l >>  8) & 0xFF] & "\x04\x04\x04\x04\x04\x04\x04\x04") |
-                        ($shuffleip[($l >> 16) & 0xFF] & "\x02\x02\x02\x02\x02\x02\x02\x02") |
-                        ($shuffleip[($l >> 24) & 0xFF] & "\x01\x01\x01\x01\x01\x01\x01\x01")
-                    );
-                    ' . /* Extract L0 and R0 */ '
-                    $l = $in[1];
-                    $r = $in[2];
-                ';
+            // Perform DES or 3DES.
+            for ($ki = -1, $des_round = 0; $des_round < $des_rounds; ++$des_round) {
+                // Perform the 16 steps.
+                for ($i = 0; $i < 16; ++$i) {
+                    // start of "the Feistel (F) function" - see the following URL:
+                    // http://en.wikipedia.org/wiki/Image:Data_Encryption_Standard_InfoBox_Diagram.png
+                    // Merge key schedule.
+                    $crypt_block[$c] .= '
+                        $b1 = ((' . $r . ' >>  3) & 0x1FFFFFFF)  ^ (' . $r . ' << 29) ^ ' . $k[$c][++$ki] . ';
+                        $b2 = ((' . $r . ' >> 31) & 0x00000001)  ^ (' . $r . ' <<  1) ^ ' . $k[$c][++$ki] . ';' .
+                        /* S-box indexing. */
+                        $l . ' = $sbox1[($b1 >> 24) & 0x3F] ^ $sbox2[($b2 >> 24) & 0x3F] ^
+                                 $sbox3[($b1 >> 16) & 0x3F] ^ $sbox4[($b2 >> 16) & 0x3F] ^
+                                 $sbox5[($b1 >>  8) & 0x3F] ^ $sbox6[($b2 >>  8) & 0x3F] ^
+                                 $sbox7[ $b1        & 0x3F] ^ $sbox8[ $b2        & 0x3F] ^ ' . $l . ';
+                    ';
+                    // end of "the Feistel (F) function"
 
-                $l = '$l';
-                $r = '$r';
-
-                // Perform DES or 3DES.
-                for ($ki = -1, $des_round = 0; $des_round < $des_rounds; ++$des_round) {
-                    // Perform the 16 steps.
-                    for ($i = 0; $i < 16; ++$i) {
-                        // start of "the Feistel (F) function" - see the following URL:
-                        // http://en.wikipedia.org/wiki/Image:Data_Encryption_Standard_InfoBox_Diagram.png
-                        // Merge key schedule.
-                        $crypt_block[$c].= '
-                            $b1 = ((' . $r . ' >>  3) & 0x1FFFFFFF)  ^ (' . $r . ' << 29) ^ ' . $k[$c][++$ki] . ';
-                            $b2 = ((' . $r . ' >> 31) & 0x00000001)  ^ (' . $r . ' <<  1) ^ ' . $k[$c][++$ki] . ';' .
-                            /* S-box indexing. */
-                            $l . ' = $sbox1[($b1 >> 24) & 0x3F] ^ $sbox2[($b2 >> 24) & 0x3F] ^
-                                     $sbox3[($b1 >> 16) & 0x3F] ^ $sbox4[($b2 >> 16) & 0x3F] ^
-                                     $sbox5[($b1 >>  8) & 0x3F] ^ $sbox6[($b2 >>  8) & 0x3F] ^
-                                     $sbox7[ $b1        & 0x3F] ^ $sbox8[ $b2        & 0x3F] ^ ' . $l . ';
-                        ';
-                        // end of "the Feistel (F) function"
-
-                        // swap L & R
-                        list($l, $r) = array($r, $l);
-                    }
-                    list($l, $r) = array($r, $l);
+                    // swap L & R
+                    list($l, $r) = [$r, $l];
                 }
-
-                // Perform the inverse IP permutation.
-                $crypt_block[$c].= '$in =
-                    ($shuffleinvip[($l >> 24) & 0xFF] & "\x80\x80\x80\x80\x80\x80\x80\x80") |
-                    ($shuffleinvip[($r >> 24) & 0xFF] & "\x40\x40\x40\x40\x40\x40\x40\x40") |
-                    ($shuffleinvip[($l >> 16) & 0xFF] & "\x20\x20\x20\x20\x20\x20\x20\x20") |
-                    ($shuffleinvip[($r >> 16) & 0xFF] & "\x10\x10\x10\x10\x10\x10\x10\x10") |
-                    ($shuffleinvip[($l >>  8) & 0xFF] & "\x08\x08\x08\x08\x08\x08\x08\x08") |
-                    ($shuffleinvip[($r >>  8) & 0xFF] & "\x04\x04\x04\x04\x04\x04\x04\x04") |
-                    ($shuffleinvip[ $l        & 0xFF] & "\x02\x02\x02\x02\x02\x02\x02\x02") |
-                    ($shuffleinvip[ $r        & 0xFF] & "\x01\x01\x01\x01\x01\x01\x01\x01");
-                ';
+                list($l, $r) = [$r, $l];
             }
 
-            // Creates the inline-crypt function
-            $lambda_functions[$code_hash] = $this->_createInlineCryptFunction(
-                array(
-                   'init_crypt'    => $init_crypt,
-                   'init_encrypt'  => $init_encrypt,
-                   'init_decrypt'  => $init_decrypt,
-                   'encrypt_block' => $crypt_block[self::ENCRYPT],
-                   'decrypt_block' => $crypt_block[self::DECRYPT]
-                )
-            );
+            // Perform the inverse IP permutation.
+            $crypt_block[$c] .= '$in =
+                ($shuffleinvip[($l >> 24) & 0xFF] & "\x80\x80\x80\x80\x80\x80\x80\x80") |
+                ($shuffleinvip[($r >> 24) & 0xFF] & "\x40\x40\x40\x40\x40\x40\x40\x40") |
+                ($shuffleinvip[($l >> 16) & 0xFF] & "\x20\x20\x20\x20\x20\x20\x20\x20") |
+                ($shuffleinvip[($r >> 16) & 0xFF] & "\x10\x10\x10\x10\x10\x10\x10\x10") |
+                ($shuffleinvip[($l >>  8) & 0xFF] & "\x08\x08\x08\x08\x08\x08\x08\x08") |
+                ($shuffleinvip[($r >>  8) & 0xFF] & "\x04\x04\x04\x04\x04\x04\x04\x04") |
+                ($shuffleinvip[ $l        & 0xFF] & "\x02\x02\x02\x02\x02\x02\x02\x02") |
+                ($shuffleinvip[ $r        & 0xFF] & "\x01\x01\x01\x01\x01\x01\x01\x01");
+            ';
         }
 
-        // Set the inline-crypt function as callback in: $this->inline_crypt
-        $this->inline_crypt = $lambda_functions[$code_hash];
+        // Creates the inline-crypt function
+        $this->inline_crypt = $this->createInlineCryptFunction(
+            [
+               'init_crypt'    => $init_crypt,
+               'init_encrypt'  => $init_encrypt,
+               'init_decrypt'  => $init_decrypt,
+               'encrypt_block' => $crypt_block[self::ENCRYPT],
+               'decrypt_block' => $crypt_block[self::DECRYPT]
+            ]
+        );
     }
 }
