@@ -16,7 +16,7 @@
  * <?php
  *    include 'vendor/autoload.php';
  *
- *    $twofish = new \phpseclib\Crypt\Twofish();
+ *    $twofish = new \phpseclib3\Crypt\Twofish('ctr');
  *
  *    $twofish->setKey('12345678901234567890123456789012');
  *
@@ -26,8 +26,6 @@
  * ?>
  * </code>
  *
- * @category  Crypt
- * @package   Twofish
  * @author    Jim Wigginton <terrafrost@php.net>
  * @author    Hans-Juergen Petrich <petrich@tronic-media.com>
  * @copyright 2007 Jim Wigginton
@@ -35,43 +33,41 @@
  * @link      http://phpseclib.sourceforge.net
  */
 
-namespace phpseclib\Crypt;
+namespace phpseclib3\Crypt;
+
+use phpseclib3\Crypt\Common\BlockCipher;
+use phpseclib3\Exception\BadModeException;
 
 /**
  * Pure-PHP implementation of Twofish.
  *
- * @package Twofish
  * @author  Jim Wigginton <terrafrost@php.net>
  * @author  Hans-Juergen Petrich <petrich@tronic-media.com>
- * @access  public
  */
-class Twofish extends Base
+class Twofish extends BlockCipher
 {
     /**
      * The mcrypt specific name of the cipher
      *
-     * @see \phpseclib\Crypt\Base::cipher_name_mcrypt
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::cipher_name_mcrypt
      * @var string
-     * @access private
      */
-    var $cipher_name_mcrypt = 'twofish';
+    protected $cipher_name_mcrypt = 'twofish';
 
     /**
      * Optimizing value while CFB-encrypting
      *
-     * @see \phpseclib\Crypt\Base::cfb_init_len
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::cfb_init_len
      * @var int
-     * @access private
      */
-    var $cfb_init_len = 800;
+    protected $cfb_init_len = 800;
 
     /**
      * Q-Table
      *
      * @var array
-     * @access private
      */
-    var $q0 = array(
+    private static $q0 = [
         0xA9, 0x67, 0xB3, 0xE8, 0x04, 0xFD, 0xA3, 0x76,
         0x9A, 0x92, 0x80, 0x78, 0xE4, 0xDD, 0xD1, 0x38,
         0x0D, 0xC6, 0x35, 0x98, 0x18, 0xF7, 0xEC, 0x6C,
@@ -104,15 +100,14 @@ class Twofish extends Base
         0xC8, 0xA8, 0x2B, 0x40, 0xDC, 0xFE, 0x32, 0xA4,
         0xCA, 0x10, 0x21, 0xF0, 0xD3, 0x5D, 0x0F, 0x00,
         0x6F, 0x9D, 0x36, 0x42, 0x4A, 0x5E, 0xC1, 0xE0
-    );
+    ];
 
     /**
      * Q-Table
      *
      * @var array
-     * @access private
      */
-    var $q1 = array(
+    private static $q1 = [
         0x75, 0xF3, 0xC6, 0xF4, 0xDB, 0x7B, 0xFB, 0xC8,
         0x4A, 0xD3, 0xE6, 0x6B, 0x45, 0x7D, 0xE8, 0x4B,
         0xD6, 0x32, 0xD8, 0xFD, 0x37, 0x71, 0xF1, 0xE1,
@@ -145,15 +140,14 @@ class Twofish extends Base
         0x12, 0xA2, 0x0D, 0x52, 0xBB, 0x02, 0x2F, 0xA9,
         0xD7, 0x61, 0x1E, 0xB4, 0x50, 0x04, 0xF6, 0xC2,
         0x16, 0x25, 0x86, 0x56, 0x55, 0x09, 0xBE, 0x91
-    );
+    ];
 
     /**
      * M-Table
      *
      * @var array
-     * @access private
      */
-    var $m0 = array(
+    private static $m0 = [
         0xBCBC3275, 0xECEC21F3, 0x202043C6, 0xB3B3C9F4, 0xDADA03DB, 0x02028B7B, 0xE2E22BFB, 0x9E9EFAC8,
         0xC9C9EC4A, 0xD4D409D3, 0x18186BE6, 0x1E1E9F6B, 0x98980E45, 0xB2B2387D, 0xA6A6D2E8, 0x2626B74B,
         0x3C3C57D6, 0x93938A32, 0x8282EED8, 0x525298FD, 0x7B7BD437, 0xBBBB3771, 0x5B5B97F1, 0x474783E1,
@@ -186,15 +180,14 @@ class Twofish extends Base
         0xABABA212, 0x6F6F3EA2, 0xE6E6540D, 0xDBDBF252, 0x92927BBB, 0xB7B7B602, 0x6969CA2F, 0x3939D9A9,
         0xD3D30CD7, 0xA7A72361, 0xA2A2AD1E, 0xC3C399B4, 0x6C6C4450, 0x07070504, 0x04047FF6, 0x272746C2,
         0xACACA716, 0xD0D07625, 0x50501386, 0xDCDCF756, 0x84841A55, 0xE1E15109, 0x7A7A25BE, 0x1313EF91
-    );
+    ];
 
     /**
      * M-Table
      *
      * @var array
-     * @access private
      */
-    var $m1 = array(
+    private static $m1 = [
         0xA9D93939, 0x67901717, 0xB3719C9C, 0xE8D2A6A6, 0x04050707, 0xFD985252, 0xA3658080, 0x76DFE4E4,
         0x9A084545, 0x92024B4B, 0x80A0E0E0, 0x78665A5A, 0xE4DDAFAF, 0xDDB06A6A, 0xD1BF6363, 0x38362A2A,
         0x0D54E6E6, 0xC6432020, 0x3562CCCC, 0x98BEF2F2, 0x181E1212, 0xF724EBEB, 0xECD7A1A1, 0x6C774141,
@@ -227,15 +220,14 @@ class Twofish extends Base
         0xC8FA9E9E, 0xA882D6D6, 0x2BCF6E6E, 0x40507070, 0xDCEB8585, 0xFE750A0A, 0x328A9393, 0xA48DDFDF,
         0xCA4C2929, 0x10141C1C, 0x2173D7D7, 0xF0CCB4B4, 0xD309D4D4, 0x5D108A8A, 0x0FE25151, 0x00000000,
         0x6F9A1919, 0x9DE01A1A, 0x368F9494, 0x42E6C7C7, 0x4AECC9C9, 0x5EFDD2D2, 0xC1AB7F7F, 0xE0D8A8A8
-    );
+    ];
 
     /**
      * M-Table
      *
      * @var array
-     * @access private
      */
-    var $m2 = array(
+    private static $m2 = [
         0xBC75BC32, 0xECF3EC21, 0x20C62043, 0xB3F4B3C9, 0xDADBDA03, 0x027B028B, 0xE2FBE22B, 0x9EC89EFA,
         0xC94AC9EC, 0xD4D3D409, 0x18E6186B, 0x1E6B1E9F, 0x9845980E, 0xB27DB238, 0xA6E8A6D2, 0x264B26B7,
         0x3CD63C57, 0x9332938A, 0x82D882EE, 0x52FD5298, 0x7B377BD4, 0xBB71BB37, 0x5BF15B97, 0x47E14783,
@@ -268,15 +260,14 @@ class Twofish extends Base
         0xAB12ABA2, 0x6FA26F3E, 0xE60DE654, 0xDB52DBF2, 0x92BB927B, 0xB702B7B6, 0x692F69CA, 0x39A939D9,
         0xD3D7D30C, 0xA761A723, 0xA21EA2AD, 0xC3B4C399, 0x6C506C44, 0x07040705, 0x04F6047F, 0x27C22746,
         0xAC16ACA7, 0xD025D076, 0x50865013, 0xDC56DCF7, 0x8455841A, 0xE109E151, 0x7ABE7A25, 0x139113EF
-    );
+    ];
 
     /**
      * M-Table
      *
      * @var array
-     * @access private
      */
-    var $m3 = array(
+    private static $m3 = [
         0xD939A9D9, 0x90176790, 0x719CB371, 0xD2A6E8D2, 0x05070405, 0x9852FD98, 0x6580A365, 0xDFE476DF,
         0x08459A08, 0x024B9202, 0xA0E080A0, 0x665A7866, 0xDDAFE4DD, 0xB06ADDB0, 0xBF63D1BF, 0x362A3836,
         0x54E60D54, 0x4320C643, 0x62CC3562, 0xBEF298BE, 0x1E12181E, 0x24EBF724, 0xD7A1ECD7, 0x77416C77,
@@ -309,120 +300,164 @@ class Twofish extends Base
         0xFA9EC8FA, 0x82D6A882, 0xCF6E2BCF, 0x50704050, 0xEB85DCEB, 0x750AFE75, 0x8A93328A, 0x8DDFA48D,
         0x4C29CA4C, 0x141C1014, 0x73D72173, 0xCCB4F0CC, 0x09D4D309, 0x108A5D10, 0xE2510FE2, 0x00000000,
         0x9A196F9A, 0xE01A9DE0, 0x8F94368F, 0xE6C742E6, 0xECC94AEC, 0xFDD25EFD, 0xAB7FC1AB, 0xD8A8E0D8
-    );
+    ];
 
     /**
      * The Key Schedule Array
      *
      * @var array
-     * @access private
      */
-    var $K = array();
+    private $K = [];
 
     /**
      * The Key depended S-Table 0
      *
      * @var array
-     * @access private
      */
-    var $S0 = array();
+    private $S0 = [];
 
     /**
      * The Key depended S-Table 1
      *
      * @var array
-     * @access private
      */
-    var $S1 = array();
+    private $S1 = [];
 
     /**
      * The Key depended S-Table 2
      *
      * @var array
-     * @access private
      */
-    var $S2 = array();
+    private $S2 = [];
 
     /**
      * The Key depended S-Table 3
      *
      * @var array
-     * @access private
      */
-    var $S3 = array();
+    private $S3 = [];
 
     /**
      * Holds the last used key
      *
      * @var array
-     * @access private
      */
-    var $kl;
+    private $kl;
 
     /**
      * The Key Length (in bytes)
      *
      * @see Crypt_Twofish::setKeyLength()
      * @var int
-     * @access private
      */
-    var $key_length = 16;
+    protected $key_length = 16;
+
+    /**
+     * Default Constructor.
+     *
+     * @param string $mode
+     * @throws BadModeException if an invalid / unsupported mode is provided
+     */
+    public function __construct($mode)
+    {
+        parent::__construct($mode);
+
+        if ($this->mode == self::MODE_STREAM) {
+            throw new BadModeException('Block ciphers cannot be ran in stream mode');
+        }
+    }
+
+    /**
+     * Initialize Static Variables
+     */
+    protected static function initialize_static_variables()
+    {
+        if (is_float(self::$m3[0])) {
+            self::$m0 = array_map('intval', self::$m0);
+            self::$m1 = array_map('intval', self::$m1);
+            self::$m2 = array_map('intval', self::$m2);
+            self::$m3 = array_map('intval', self::$m3);
+            self::$q0 = array_map('intval', self::$q0);
+            self::$q1 = array_map('intval', self::$q1);
+        }
+
+        parent::initialize_static_variables();
+    }
 
     /**
      * Sets the key length.
      *
      * Valid key lengths are 128, 192 or 256 bits
      *
-     * @access public
      * @param int $length
      */
-    function setKeyLength($length)
+    public function setKeyLength($length)
     {
-        switch (true) {
-            case $length <= 128:
-                $this->key_length = 16;
-                break;
-            case $length <= 192:
-                $this->key_length = 24;
+        switch ($length) {
+            case 128:
+            case 192:
+            case 256:
                 break;
             default:
-                $this->key_length = 32;
+                throw new \LengthException('Key of size ' . $length . ' not supported by this algorithm. Only keys of sizes 16, 24 or 32 supported');
         }
 
         parent::setKeyLength($length);
     }
 
     /**
+     * Sets the key.
+     *
+     * Rijndael supports five different key lengths
+     *
+     * @see setKeyLength()
+     * @param string $key
+     * @throws \LengthException if the key length isn't supported
+     */
+    public function setKey($key)
+    {
+        switch (strlen($key)) {
+            case 16:
+            case 24:
+            case 32:
+                break;
+            default:
+                throw new \LengthException('Key of size ' . strlen($key) . ' not supported by this algorithm. Only keys of sizes 16, 24 or 32 supported');
+        }
+
+        parent::setKey($key);
+    }
+
+    /**
      * Setup the key (expansion)
      *
-     * @see \phpseclib\Crypt\Base::_setupKey()
-     * @access private
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::_setupKey()
      */
-    function _setupKey()
+    protected function setupKey()
     {
         if (isset($this->kl['key']) && $this->key === $this->kl['key']) {
             // already expanded
             return;
         }
-        $this->kl = array('key' => $this->key);
+        $this->kl = ['key' => $this->key];
 
         /* Key expanding and generating the key-depended s-boxes */
         $le_longs = unpack('V*', $this->key);
         $key = unpack('C*', $this->key);
-        $m0 = $this->m0;
-        $m1 = $this->m1;
-        $m2 = $this->m2;
-        $m3 = $this->m3;
-        $q0 = $this->q0;
-        $q1 = $this->q1;
+        $m0 = self::$m0;
+        $m1 = self::$m1;
+        $m2 = self::$m2;
+        $m3 = self::$m3;
+        $q0 = self::$q0;
+        $q1 = self::$q1;
 
-        $K = $S0 = $S1 = $S2 = $S3 = array();
+        $K = $S0 = $S1 = $S2 = $S3 = [];
 
         switch (strlen($this->key)) {
             case 16:
-                list($s7, $s6, $s5, $s4) = $this->_mdsrem($le_longs[1], $le_longs[2]);
-                list($s3, $s2, $s1, $s0) = $this->_mdsrem($le_longs[3], $le_longs[4]);
-                for ($i = 0, $j = 1; $i < 40; $i+= 2, $j+= 2) {
+                list($s7, $s6, $s5, $s4) = $this->mdsrem($le_longs[1], $le_longs[2]);
+                list($s3, $s2, $s1, $s0) = $this->mdsrem($le_longs[3], $le_longs[4]);
+                for ($i = 0, $j = 1; $i < 40; $i += 2, $j += 2) {
                     $A = $m0[$q0[$q0[$i] ^ $key[ 9]] ^ $key[1]] ^
                          $m1[$q0[$q1[$i] ^ $key[10]] ^ $key[2]] ^
                          $m2[$q1[$q0[$i] ^ $key[11]] ^ $key[3]] ^
@@ -432,9 +467,9 @@ class Twofish extends Base
                          $m2[$q1[$q0[$j] ^ $key[15]] ^ $key[7]] ^
                          $m3[$q1[$q1[$j] ^ $key[16]] ^ $key[8]];
                     $B = ($B << 8) | ($B >> 24 & 0xff);
-                    $A = $this->safe_intval($A + $B);
+                    $A = self::safe_intval($A + $B);
                     $K[] = $A;
-                    $A = $this->safe_intval($A + $B);
+                    $A = self::safe_intval($A + $B);
                     $K[] = ($A << 9 | $A >> 23 & 0x1ff);
                 }
                 for ($i = 0; $i < 256; ++$i) {
@@ -445,10 +480,10 @@ class Twofish extends Base
                 }
                 break;
             case 24:
-                list($sb, $sa, $s9, $s8) = $this->_mdsrem($le_longs[1], $le_longs[2]);
-                list($s7, $s6, $s5, $s4) = $this->_mdsrem($le_longs[3], $le_longs[4]);
-                list($s3, $s2, $s1, $s0) = $this->_mdsrem($le_longs[5], $le_longs[6]);
-                for ($i = 0, $j = 1; $i < 40; $i+= 2, $j+= 2) {
+                list($sb, $sa, $s9, $s8) = $this->mdsrem($le_longs[1], $le_longs[2]);
+                list($s7, $s6, $s5, $s4) = $this->mdsrem($le_longs[3], $le_longs[4]);
+                list($s3, $s2, $s1, $s0) = $this->mdsrem($le_longs[5], $le_longs[6]);
+                for ($i = 0, $j = 1; $i < 40; $i += 2, $j += 2) {
                     $A = $m0[$q0[$q0[$q1[$i] ^ $key[17]] ^ $key[ 9]] ^ $key[1]] ^
                          $m1[$q0[$q1[$q1[$i] ^ $key[18]] ^ $key[10]] ^ $key[2]] ^
                          $m2[$q1[$q0[$q0[$i] ^ $key[19]] ^ $key[11]] ^ $key[3]] ^
@@ -458,9 +493,9 @@ class Twofish extends Base
                          $m2[$q1[$q0[$q0[$j] ^ $key[23]] ^ $key[15]] ^ $key[7]] ^
                          $m3[$q1[$q1[$q0[$j] ^ $key[24]] ^ $key[16]] ^ $key[8]];
                     $B = ($B << 8) | ($B >> 24 & 0xff);
-                    $A = $this->safe_intval($A + $B);
+                    $A = self::safe_intval($A + $B);
                     $K[] = $A;
-                    $A = $this->safe_intval($A + $B);
+                    $A = self::safe_intval($A + $B);
                     $K[] = ($A << 9 | $A >> 23 & 0x1ff);
                 }
                 for ($i = 0; $i < 256; ++$i) {
@@ -471,11 +506,11 @@ class Twofish extends Base
                 }
                 break;
             default: // 32
-                list($sf, $se, $sd, $sc) = $this->_mdsrem($le_longs[1], $le_longs[2]);
-                list($sb, $sa, $s9, $s8) = $this->_mdsrem($le_longs[3], $le_longs[4]);
-                list($s7, $s6, $s5, $s4) = $this->_mdsrem($le_longs[5], $le_longs[6]);
-                list($s3, $s2, $s1, $s0) = $this->_mdsrem($le_longs[7], $le_longs[8]);
-                for ($i = 0, $j = 1; $i < 40; $i+= 2, $j+= 2) {
+                list($sf, $se, $sd, $sc) = $this->mdsrem($le_longs[1], $le_longs[2]);
+                list($sb, $sa, $s9, $s8) = $this->mdsrem($le_longs[3], $le_longs[4]);
+                list($s7, $s6, $s5, $s4) = $this->mdsrem($le_longs[5], $le_longs[6]);
+                list($s3, $s2, $s1, $s0) = $this->mdsrem($le_longs[7], $le_longs[8]);
+                for ($i = 0, $j = 1; $i < 40; $i += 2, $j += 2) {
                     $A = $m0[$q0[$q0[$q1[$q1[$i] ^ $key[25]] ^ $key[17]] ^ $key[ 9]] ^ $key[1]] ^
                          $m1[$q0[$q1[$q1[$q0[$i] ^ $key[26]] ^ $key[18]] ^ $key[10]] ^ $key[2]] ^
                          $m2[$q1[$q0[$q0[$q0[$i] ^ $key[27]] ^ $key[19]] ^ $key[11]] ^ $key[3]] ^
@@ -485,9 +520,9 @@ class Twofish extends Base
                          $m2[$q1[$q0[$q0[$q0[$j] ^ $key[31]] ^ $key[23]] ^ $key[15]] ^ $key[7]] ^
                          $m3[$q1[$q1[$q0[$q1[$j] ^ $key[32]] ^ $key[24]] ^ $key[16]] ^ $key[8]];
                     $B = ($B << 8) | ($B >> 24 & 0xff);
-                    $A = $this->safe_intval($A + $B);
+                    $A = self::safe_intval($A + $B);
                     $K[] = $A;
-                    $A = $this->safe_intval($A + $B);
+                    $A = self::safe_intval($A + $B);
                     $K[] = ($A << 9 | $A >> 23 & 0x1ff);
                 }
                 for ($i = 0; $i < 256; ++$i) {
@@ -508,12 +543,11 @@ class Twofish extends Base
     /**
      * _mdsrem function using by the twofish cipher algorithm
      *
-     * @access private
      * @param string $A
      * @param string $B
      * @return array
      */
-    function _mdsrem($A, $B)
+    private function mdsrem($A, $B)
     {
         // No gain by unrolling this loop.
         for ($i = 0; $i < 8; ++$i) {
@@ -522,45 +556,44 @@ class Twofish extends Base
 
             // Shift the others up.
             $B = ($B << 8) | (0xff & ($A >> 24));
-            $A<<= 8;
+            $A <<= 8;
 
             $u = $t << 1;
 
             // Subtract the modular polynomial on overflow.
             if ($t & 0x80) {
-                $u^= 0x14d;
+                $u ^= 0x14d;
             }
 
             // Remove t * (a * x^2 + 1).
             $B ^= $t ^ ($u << 16);
 
             // Form u = a*t + t/a = t*(a + 1/a).
-            $u^= 0x7fffffff & ($t >> 1);
+            $u ^= 0x7fffffff & ($t >> 1);
 
             // Add the modular polynomial on underflow.
             if ($t & 0x01) {
-                $u^= 0xa6 ;
+                $u ^= 0xa6 ;
             }
 
             // Remove t * (a + 1/a) * (x^3 + x).
-            $B^= ($u << 24) | ($u << 8);
+            $B ^= ($u << 24) | ($u << 8);
         }
 
-        return array(
+        return [
             0xff & $B >> 24,
             0xff & $B >> 16,
             0xff & $B >>  8,
-            0xff & $B);
+            0xff & $B];
     }
 
     /**
      * Encrypts a block
      *
-     * @access private
      * @param string $in
      * @return string
      */
-    function _encryptBlock($in)
+    protected function encryptBlock($in)
     {
         $S0 = $this->S0;
         $S1 = $this->S1;
@@ -584,9 +617,9 @@ class Twofish extends Base
                   $S1[ $R1        & 0xff] ^
                   $S2[($R1 >>  8) & 0xff] ^
                   $S3[($R1 >> 16) & 0xff];
-            $R2^= $this->safe_intval($t0 + $t1 + $K[++$ki]);
+            $R2 ^= self::safe_intval($t0 + $t1 + $K[++$ki]);
             $R2 = ($R2 >> 1 & 0x7fffffff) | ($R2 << 31);
-            $R3 = ((($R3 >> 31) & 1) | ($R3 << 1)) ^ $this->safe_intval($t0 + ($t1 << 1) + $K[++$ki]);
+            $R3 = ((($R3 >> 31) & 1) | ($R3 << 1)) ^ self::safe_intval($t0 + ($t1 << 1) + $K[++$ki]);
 
             $t0 = $S0[ $R2        & 0xff] ^
                   $S1[($R2 >>  8) & 0xff] ^
@@ -596,9 +629,9 @@ class Twofish extends Base
                   $S1[ $R3        & 0xff] ^
                   $S2[($R3 >>  8) & 0xff] ^
                   $S3[($R3 >> 16) & 0xff];
-            $R0^= $this->safe_intval($t0 + $t1 + $K[++$ki]);
+            $R0 ^= self::safe_intval($t0 + $t1 + $K[++$ki]);
             $R0 = ($R0 >> 1 & 0x7fffffff) | ($R0 << 31);
-            $R1 = ((($R1 >> 31) & 1) | ($R1 << 1)) ^ $this->safe_intval($t0 + ($t1 << 1) + $K[++$ki]);
+            $R1 = ((($R1 >> 31) & 1) | ($R1 << 1)) ^ self::safe_intval($t0 + ($t1 << 1) + $K[++$ki]);
         }
 
         // @codingStandardsIgnoreStart
@@ -612,11 +645,10 @@ class Twofish extends Base
     /**
      * Decrypts a block
      *
-     * @access private
      * @param string $in
      * @return string
      */
-    function _decryptBlock($in)
+    protected function decryptBlock($in)
     {
         $S0 = $this->S0;
         $S1 = $this->S1;
@@ -640,9 +672,9 @@ class Twofish extends Base
                   $S1[$R1       & 0xff] ^
                   $S2[$R1 >>  8 & 0xff] ^
                   $S3[$R1 >> 16 & 0xff];
-            $R3^= $this->safe_intval($t0 + ($t1 << 1) + $K[--$ki]);
+            $R3 ^= self::safe_intval($t0 + ($t1 << 1) + $K[--$ki]);
             $R3 = $R3 >> 1 & 0x7fffffff | $R3 << 31;
-            $R2 = ($R2 >> 31 & 0x1 | $R2 << 1) ^ $this->safe_intval($t0 + $t1 + $K[--$ki]);
+            $R2 = ($R2 >> 31 & 0x1 | $R2 << 1) ^ self::safe_intval($t0 + $t1 + $K[--$ki]);
 
             $t0 = $S0[$R2       & 0xff] ^
                   $S1[$R2 >>  8 & 0xff] ^
@@ -652,9 +684,9 @@ class Twofish extends Base
                   $S1[$R3       & 0xff] ^
                   $S2[$R3 >>  8 & 0xff] ^
                   $S3[$R3 >> 16 & 0xff];
-            $R1^= $this->safe_intval($t0 + ($t1 << 1) + $K[--$ki]);
+            $R1 ^= self::safe_intval($t0 + ($t1 << 1) + $K[--$ki]);
             $R1 = $R1 >> 1 & 0x7fffffff | $R1 << 31;
-            $R0 = ($R0 >> 31 & 0x1 | $R0 << 1) ^ $this->safe_intval($t0 + $t1 + $K[--$ki]);
+            $R0 = ($R0 >> 31 & 0x1 | $R0 << 1) ^ self::safe_intval($t0 + $t1 + $K[--$ki]);
         }
 
         // @codingStandardsIgnoreStart
@@ -668,149 +700,117 @@ class Twofish extends Base
     /**
      * Setup the performance-optimized function for de/encrypt()
      *
-     * @see \phpseclib\Crypt\Base::_setupInlineCrypt()
-     * @access private
+     * @see \phpseclib3\Crypt\Common\SymmetricKey::_setupInlineCrypt()
      */
-    function _setupInlineCrypt()
+    protected function setupInlineCrypt()
     {
-        $lambda_functions =& self::_getLambdaFunctions();
-
-        // Max. 10 Ultra-Hi-optimized inline-crypt functions. After that, we'll (still) create very fast code, but not the ultimate fast one.
-        // (Currently, for Crypt_Twofish, one generated $lambda_function cost on php5.5@32bit ~140kb unfreeable mem and ~240kb on php5.5@64bit)
-        $gen_hi_opt_code = (bool)(count($lambda_functions) < 10);
-
-        // Generation of a unique hash for our generated code
-        $code_hash = "Crypt_Twofish, {$this->mode}";
-        if ($gen_hi_opt_code) {
-            $code_hash = str_pad($code_hash, 32) . $this->_hashInlineCryptFunction($this->key);
-        }
-
-        $safeint = $this->safe_intval_inline();
-
-        if (!isset($lambda_functions[$code_hash])) {
-            switch (true) {
-                case $gen_hi_opt_code:
-                    $K = $this->K;
-                    $init_crypt = '
-                        static $S0, $S1, $S2, $S3;
-                        if (!$S0) {
-                            for ($i = 0; $i < 256; ++$i) {
-                                $S0[] = (int)$self->S0[$i];
-                                $S1[] = (int)$self->S1[$i];
-                                $S2[] = (int)$self->S2[$i];
-                                $S3[] = (int)$self->S3[$i];
-                            }
-                        }
-                    ';
-                    break;
-                default:
-                    $K   = array();
-                    for ($i = 0; $i < 40; ++$i) {
-                        $K[] = '$K_' . $i;
-                    }
-                    $init_crypt = '
-                        $S0 = $self->S0;
-                        $S1 = $self->S1;
-                        $S2 = $self->S2;
-                        $S3 = $self->S3;
-                        list(' . implode(',', $K) . ') = $self->K;
-                    ';
+        $K = $this->K;
+        $init_crypt = '
+            static $S0, $S1, $S2, $S3;
+            if (!$S0) {
+                for ($i = 0; $i < 256; ++$i) {
+                    $S0[] = (int)$this->S0[$i];
+                    $S1[] = (int)$this->S1[$i];
+                    $S2[] = (int)$this->S2[$i];
+                    $S3[] = (int)$this->S3[$i];
+                }
             }
+        ';
 
-            // Generating encrypt code:
-            $encrypt_block = '
-                $in = unpack("V4", $in);
-                $R0 = '.$K[0].' ^ $in[1];
-                $R1 = '.$K[1].' ^ $in[2];
-                $R2 = '.$K[2].' ^ $in[3];
-                $R3 = '.$K[3].' ^ $in[4];
-            ';
-            for ($ki = 7, $i = 0; $i < 8; ++$i) {
-                $encrypt_block.= '
-                    $t0 = $S0[ $R0        & 0xff] ^
-                          $S1[($R0 >>  8) & 0xff] ^
-                          $S2[($R0 >> 16) & 0xff] ^
-                          $S3[($R0 >> 24) & 0xff];
-                    $t1 = $S0[($R1 >> 24) & 0xff] ^
-                          $S1[ $R1        & 0xff] ^
-                          $S2[($R1 >>  8) & 0xff] ^
-                          $S3[($R1 >> 16) & 0xff];
+        $safeint = self::safe_intval_inline();
+
+        // Generating encrypt code:
+        $encrypt_block = '
+            $in = unpack("V4", $in);
+            $R0 = ' . $K[0] . ' ^ $in[1];
+            $R1 = ' . $K[1] . ' ^ $in[2];
+            $R2 = ' . $K[2] . ' ^ $in[3];
+            $R3 = ' . $K[3] . ' ^ $in[4];
+        ';
+        for ($ki = 7, $i = 0; $i < 8; ++$i) {
+            $encrypt_block .= '
+                $t0 = $S0[ $R0        & 0xff] ^
+                      $S1[($R0 >>  8) & 0xff] ^
+                      $S2[($R0 >> 16) & 0xff] ^
+                      $S3[($R0 >> 24) & 0xff];
+                $t1 = $S0[($R1 >> 24) & 0xff] ^
+                      $S1[ $R1        & 0xff] ^
+                      $S2[($R1 >>  8) & 0xff] ^
+                      $S3[($R1 >> 16) & 0xff];
                     $R2^= ' . sprintf($safeint, '$t0 + $t1 + ' . $K[++$ki]) . ';
-                    $R2 = ($R2 >> 1 & 0x7fffffff) | ($R2 << 31);
-                    $R3 = ((($R3 >> 31) & 1) | ($R3 << 1)) ^ ' . sprintf($safeint, '($t0 + ($t1 << 1) + ' . $K[++$ki] . ')') . ';
+                $R2 = ($R2 >> 1 & 0x7fffffff) | ($R2 << 31);
+                $R3 = ((($R3 >> 31) & 1) | ($R3 << 1)) ^ ' . sprintf($safeint, '($t0 + ($t1 << 1) + ' . $K[++$ki] . ')') . ';
 
-                    $t0 = $S0[ $R2        & 0xff] ^
-                          $S1[($R2 >>  8) & 0xff] ^
-                          $S2[($R2 >> 16) & 0xff] ^
-                          $S3[($R2 >> 24) & 0xff];
-                    $t1 = $S0[($R3 >> 24) & 0xff] ^
-                          $S1[ $R3        & 0xff] ^
-                          $S2[($R3 >>  8) & 0xff] ^
-                          $S3[($R3 >> 16) & 0xff];
-                    $R0^= ' . sprintf($safeint, '($t0 + $t1 + ' . $K[++$ki] . ')') . ';
-                    $R0 = ($R0 >> 1 & 0x7fffffff) | ($R0 << 31);
-                    $R1 = ((($R1 >> 31) & 1) | ($R1 << 1)) ^ ' . sprintf($safeint, '($t0 + ($t1 << 1) + ' . $K[++$ki] . ')') . ';
-                ';
-            }
-            $encrypt_block.= '
-                $in = pack("V4", ' . $K[4] . ' ^ $R2,
-                                 ' . $K[5] . ' ^ $R3,
-                                 ' . $K[6] . ' ^ $R0,
-                                 ' . $K[7] . ' ^ $R1);
+                $t0 = $S0[ $R2        & 0xff] ^
+                      $S1[($R2 >>  8) & 0xff] ^
+                      $S2[($R2 >> 16) & 0xff] ^
+                      $S3[($R2 >> 24) & 0xff];
+                $t1 = $S0[($R3 >> 24) & 0xff] ^
+                      $S1[ $R3        & 0xff] ^
+                      $S2[($R3 >>  8) & 0xff] ^
+                      $S3[($R3 >> 16) & 0xff];
+                $R0^= ' . sprintf($safeint, '($t0 + $t1 + ' . $K[++$ki] . ')') . ';
+                $R0 = ($R0 >> 1 & 0x7fffffff) | ($R0 << 31);
+                $R1 = ((($R1 >> 31) & 1) | ($R1 << 1)) ^ ' . sprintf($safeint, '($t0 + ($t1 << 1) + ' . $K[++$ki] . ')') . ';
             ';
-
-            // Generating decrypt code:
-            $decrypt_block = '
-                $in = unpack("V4", $in);
-                $R0 = '.$K[4].' ^ $in[1];
-                $R1 = '.$K[5].' ^ $in[2];
-                $R2 = '.$K[6].' ^ $in[3];
-                $R3 = '.$K[7].' ^ $in[4];
-            ';
-            for ($ki = 40, $i = 0; $i < 8; ++$i) {
-                $decrypt_block.= '
-                    $t0 = $S0[$R0       & 0xff] ^
-                          $S1[$R0 >>  8 & 0xff] ^
-                          $S2[$R0 >> 16 & 0xff] ^
-                          $S3[$R0 >> 24 & 0xff];
-                    $t1 = $S0[$R1 >> 24 & 0xff] ^
-                          $S1[$R1       & 0xff] ^
-                          $S2[$R1 >>  8 & 0xff] ^
-                          $S3[$R1 >> 16 & 0xff];
-                    $R3^= ' . sprintf($safeint, '$t0 + ($t1 << 1) + ' . $K[--$ki]) . ';
-                    $R3 = $R3 >> 1 & 0x7fffffff | $R3 << 31;
-                    $R2 = ($R2 >> 31 & 0x1 | $R2 << 1) ^ ' . sprintf($safeint, '($t0 + $t1 + '.$K[--$ki] . ')') . ';
-
-                    $t0 = $S0[$R2       & 0xff] ^
-                          $S1[$R2 >>  8 & 0xff] ^
-                          $S2[$R2 >> 16 & 0xff] ^
-                          $S3[$R2 >> 24 & 0xff];
-                    $t1 = $S0[$R3 >> 24 & 0xff] ^
-                          $S1[$R3       & 0xff] ^
-                          $S2[$R3 >>  8 & 0xff] ^
-                          $S3[$R3 >> 16 & 0xff];
-                    $R1^= ' . sprintf($safeint, '$t0 + ($t1 << 1) + ' . $K[--$ki]) . ';
-                    $R1 = $R1 >> 1 & 0x7fffffff | $R1 << 31;
-                    $R0 = ($R0 >> 31 & 0x1 | $R0 << 1) ^ ' . sprintf($safeint, '($t0 + $t1 + '.$K[--$ki] . ')') . ';
-                ';
-            }
-            $decrypt_block.= '
-                $in = pack("V4", ' . $K[0] . ' ^ $R2,
-                                 ' . $K[1] . ' ^ $R3,
-                                 ' . $K[2] . ' ^ $R0,
-                                 ' . $K[3] . ' ^ $R1);
-            ';
-
-            $lambda_functions[$code_hash] = $this->_createInlineCryptFunction(
-                array(
-                   'init_crypt'    => $init_crypt,
-                   'init_encrypt'  => '',
-                   'init_decrypt'  => '',
-                   'encrypt_block' => $encrypt_block,
-                   'decrypt_block' => $decrypt_block
-                )
-            );
         }
-        $this->inline_crypt = $lambda_functions[$code_hash];
+        $encrypt_block .= '
+            $in = pack("V4", ' . $K[4] . ' ^ $R2,
+                             ' . $K[5] . ' ^ $R3,
+                             ' . $K[6] . ' ^ $R0,
+                             ' . $K[7] . ' ^ $R1);
+        ';
+
+        // Generating decrypt code:
+        $decrypt_block = '
+            $in = unpack("V4", $in);
+            $R0 = ' . $K[4] . ' ^ $in[1];
+            $R1 = ' . $K[5] . ' ^ $in[2];
+            $R2 = ' . $K[6] . ' ^ $in[3];
+            $R3 = ' . $K[7] . ' ^ $in[4];
+        ';
+        for ($ki = 40, $i = 0; $i < 8; ++$i) {
+            $decrypt_block .= '
+                $t0 = $S0[$R0       & 0xff] ^
+                      $S1[$R0 >>  8 & 0xff] ^
+                      $S2[$R0 >> 16 & 0xff] ^
+                      $S3[$R0 >> 24 & 0xff];
+                $t1 = $S0[$R1 >> 24 & 0xff] ^
+                      $S1[$R1       & 0xff] ^
+                      $S2[$R1 >>  8 & 0xff] ^
+                      $S3[$R1 >> 16 & 0xff];
+                $R3^= ' . sprintf($safeint, '$t0 + ($t1 << 1) + ' . $K[--$ki]) . ';
+                $R3 = $R3 >> 1 & 0x7fffffff | $R3 << 31;
+                $R2 = ($R2 >> 31 & 0x1 | $R2 << 1) ^ ' . sprintf($safeint, '($t0 + $t1 + ' . $K[--$ki] . ')') . ';
+
+                $t0 = $S0[$R2       & 0xff] ^
+                      $S1[$R2 >>  8 & 0xff] ^
+                      $S2[$R2 >> 16 & 0xff] ^
+                      $S3[$R2 >> 24 & 0xff];
+                $t1 = $S0[$R3 >> 24 & 0xff] ^
+                      $S1[$R3       & 0xff] ^
+                      $S2[$R3 >>  8 & 0xff] ^
+                      $S3[$R3 >> 16 & 0xff];
+                $R1^= ' . sprintf($safeint, '$t0 + ($t1 << 1) + ' . $K[--$ki]) . ';
+                $R1 = $R1 >> 1 & 0x7fffffff | $R1 << 31;
+                $R0 = ($R0 >> 31 & 0x1 | $R0 << 1) ^ ' . sprintf($safeint, '($t0 + $t1 + ' . $K[--$ki] . ')') . ';
+            ';
+        }
+        $decrypt_block .= '
+            $in = pack("V4", ' . $K[0] . ' ^ $R2,
+                             ' . $K[1] . ' ^ $R3,
+                             ' . $K[2] . ' ^ $R0,
+                             ' . $K[3] . ' ^ $R1);
+        ';
+
+        $this->inline_crypt = $this->createInlineCryptFunction(
+            [
+               'init_crypt'    => $init_crypt,
+               'init_encrypt'  => '',
+               'init_decrypt'  => '',
+               'encrypt_block' => $encrypt_block,
+               'decrypt_block' => $decrypt_block
+            ]
+        );
     }
 }
