@@ -5,27 +5,23 @@
  *
  * PHP version 5
  *
- * If you call read() in \phpseclib\Net\SSH2 you may get {@link http://en.wikipedia.org/wiki/ANSI_escape_code ANSI escape codes} back.
+ * If you call read() in \phpseclib3\Net\SSH2 you may get {@link http://en.wikipedia.org/wiki/ANSI_escape_code ANSI escape codes} back.
  * They'd look like chr(0x1B) . '[00m' or whatever (0x1B = ESC).  They tell a
  * {@link http://en.wikipedia.org/wiki/Terminal_emulator terminal emulator} how to format the characters, what
- * color to display them in, etc. \phpseclib\File\ANSI is a {@link http://en.wikipedia.org/wiki/VT100 VT100} terminal emulator.
+ * color to display them in, etc. \phpseclib3\File\ANSI is a {@link http://en.wikipedia.org/wiki/VT100 VT100} terminal emulator.
  *
- * @category  File
- * @package   ANSI
  * @author    Jim Wigginton <terrafrost@php.net>
  * @copyright 2012 Jim Wigginton
  * @license   http://www.opensource.org/licenses/mit-license.html  MIT License
  * @link      http://phpseclib.sourceforge.net
  */
 
-namespace phpseclib\File;
+namespace phpseclib3\File;
 
 /**
  * Pure-PHP ANSI Decoder
  *
- * @package ANSI
  * @author  Jim Wigginton <terrafrost@php.net>
- * @access  public
  */
 class ANSI
 {
@@ -33,137 +29,120 @@ class ANSI
      * Max Width
      *
      * @var int
-     * @access private
      */
-    var $max_x;
+    private $max_x;
 
     /**
      * Max Height
      *
      * @var int
-     * @access private
      */
-    var $max_y;
+    private $max_y;
 
     /**
      * Max History
      *
      * @var int
-     * @access private
      */
-    var $max_history;
+    private $max_history;
 
     /**
      * History
      *
      * @var array
-     * @access private
      */
-    var $history;
+    private $history;
 
     /**
      * History Attributes
      *
      * @var array
-     * @access private
      */
-    var $history_attrs;
+    private $history_attrs;
 
     /**
      * Current Column
      *
      * @var int
-     * @access private
      */
-    var $x;
+    private $x;
 
     /**
      * Current Row
      *
      * @var int
-     * @access private
      */
-    var $y;
+    private $y;
 
     /**
      * Old Column
      *
      * @var int
-     * @access private
      */
-    var $old_x;
+    private $old_x;
 
     /**
      * Old Row
      *
      * @var int
-     * @access private
      */
-    var $old_y;
+    private $old_y;
 
     /**
      * An empty attribute cell
      *
      * @var object
-     * @access private
      */
-    var $base_attr_cell;
+    private $base_attr_cell;
 
     /**
      * The current attribute cell
      *
      * @var object
-     * @access private
      */
-    var $attr_cell;
+    private $attr_cell;
 
     /**
      * An empty attribute row
      *
      * @var array
-     * @access private
      */
-    var $attr_row;
+    private $attr_row;
 
     /**
      * The current screen text
      *
-     * @var array
-     * @access private
+     * @var list<string>
      */
-    var $screen;
+    private $screen;
 
     /**
      * The current screen attributes
      *
      * @var array
-     * @access private
      */
-    var $attrs;
+    private $attrs;
 
     /**
      * Current ANSI code
      *
      * @var string
-     * @access private
      */
-    var $ansi;
+    private $ansi;
 
     /**
      * Tokenization
      *
      * @var array
-     * @access private
      */
-    var $tokenization;
+    private $tokenization;
 
     /**
      * Default Constructor.
      *
-     * @return \phpseclib\File\ANSI
-     * @access public
+     * @return \phpseclib3\File\ANSI
      */
-    function __construct()
+    public function __construct()
     {
         $attr_cell = new \stdClass();
         $attr_cell->bold = false;
@@ -186,14 +165,13 @@ class ANSI
      *
      * @param int $x
      * @param int $y
-     * @access public
      */
-    function setDimensions($x, $y)
+    public function setDimensions($x, $y)
     {
         $this->max_x = $x - 1;
         $this->max_y = $y - 1;
         $this->x = $this->y = 0;
-        $this->history = $this->history_attrs = array();
+        $this->history = $this->history_attrs = [];
         $this->attr_row = array_fill(0, $this->max_x + 2, $this->base_attr_cell);
         $this->screen = array_fill(0, $this->max_y + 1, '');
         $this->attrs = array_fill(0, $this->max_y + 1, $this->attr_row);
@@ -203,11 +181,9 @@ class ANSI
     /**
      * Set the number of lines that should be logged past the terminal height
      *
-     * @param int $x
-     * @param int $y
-     * @access public
+     * @param int $history
      */
-    function setHistory($history)
+    public function setHistory($history)
     {
         $this->max_history = $history;
     }
@@ -216,9 +192,8 @@ class ANSI
      * Load a string
      *
      * @param string $source
-     * @access public
      */
-    function loadString($source)
+    public function loadString($source)
     {
         $this->setDimensions($this->max_x + 1, $this->max_y + 1);
         $this->appendString($source);
@@ -228,14 +203,13 @@ class ANSI
      * Appdend a string
      *
      * @param string $source
-     * @access public
      */
-    function appendString($source)
+    public function appendString($source)
     {
-        $this->tokenization = array('');
+        $this->tokenization = [''];
         for ($i = 0; $i < strlen($source); $i++) {
             if (strlen($this->ansi)) {
-                $this->ansi.= $source[$i];
+                $this->ansi .= $source[$i];
                 $chr = ord($source[$i]);
                 // http://en.wikipedia.org/wiki/ANSI_escape_code#Sequence_elements
                 // single character CSI's not currently supported
@@ -269,6 +243,7 @@ class ANSI
                             array_shift($this->history);
                             array_shift($this->history_attrs);
                         }
+                        // fall-through
                     case "\x1B[K": // Clear screen from cursor right
                         $this->screen[$this->y] = substr($this->screen[$this->y], 0, $this->x);
 
@@ -283,28 +258,28 @@ class ANSI
                     case "\x1B(B": // set united states g0 character set
                         break;
                     case "\x1BE": // Move to next line
-                        $this->_newLine();
+                        $this->newLine();
                         $this->x = 0;
                         break;
                     default:
                         switch (true) {
                             case preg_match('#\x1B\[(\d+)B#', $this->ansi, $match): // Move cursor down n lines
                                 $this->old_y = $this->y;
-                                $this->y+= $match[1];
+                                $this->y += (int) $match[1];
                                 break;
                             case preg_match('#\x1B\[(\d+);(\d+)H#', $this->ansi, $match): // Move cursor to screen location v,h
                                 $this->old_x = $this->x;
                                 $this->old_y = $this->y;
                                 $this->x = $match[2] - 1;
-                                $this->y = $match[1] - 1;
+                                $this->y = (int) $match[1] - 1;
                                 break;
                             case preg_match('#\x1B\[(\d+)C#', $this->ansi, $match): // Move cursor right n lines
                                 $this->old_x = $this->x;
-                                $this->x+= $match[1];
+                                $this->x += $match[1];
                                 break;
                             case preg_match('#\x1B\[(\d+)D#', $this->ansi, $match): // Move cursor left n lines
                                 $this->old_x = $this->x;
-                                $this->x-= $match[1];
+                                $this->x -= $match[1];
                                 if ($this->x < 0) {
                                     $this->x = 0;
                                 }
@@ -316,19 +291,20 @@ class ANSI
                                 $mods = explode(';', $match[1]);
                                 foreach ($mods as $mod) {
                                     switch ($mod) {
-                                        case 0: // Turn off character attributes
+                                        case '':
+                                        case '0': // Turn off character attributes
                                             $attr_cell = clone $this->base_attr_cell;
                                             break;
-                                        case 1: // Turn bold mode on
+                                        case '1': // Turn bold mode on
                                             $attr_cell->bold = true;
                                             break;
-                                        case 4: // Turn underline mode on
+                                        case '4': // Turn underline mode on
                                             $attr_cell->underline = true;
                                             break;
-                                        case 5: // Turn blinking mode on
+                                        case '5': // Turn blinking mode on
                                             $attr_cell->blink = true;
                                             break;
-                                        case 7: // Turn reverse video on
+                                        case '7': // Turn reverse video on
                                             $attr_cell->reverse = !$attr_cell->reverse;
                                             $temp = $attr_cell->background;
                                             $attr_cell->background = $attr_cell->foreground;
@@ -341,23 +317,23 @@ class ANSI
                                             $back = &$attr_cell->{ $attr_cell->reverse ? 'foreground' : 'background' };
                                             switch ($mod) {
                                                 // @codingStandardsIgnoreStart
-                                                case 30: $front = 'black'; break;
-                                                case 31: $front = 'red'; break;
-                                                case 32: $front = 'green'; break;
-                                                case 33: $front = 'yellow'; break;
-                                                case 34: $front = 'blue'; break;
-                                                case 35: $front = 'magenta'; break;
-                                                case 36: $front = 'cyan'; break;
-                                                case 37: $front = 'white'; break;
+                                                case '30': $front = 'black'; break;
+                                                case '31': $front = 'red'; break;
+                                                case '32': $front = 'green'; break;
+                                                case '33': $front = 'yellow'; break;
+                                                case '34': $front = 'blue'; break;
+                                                case '35': $front = 'magenta'; break;
+                                                case '36': $front = 'cyan'; break;
+                                                case '37': $front = 'white'; break;
 
-                                                case 40: $back = 'black'; break;
-                                                case 41: $back = 'red'; break;
-                                                case 42: $back = 'green'; break;
-                                                case 43: $back = 'yellow'; break;
-                                                case 44: $back = 'blue'; break;
-                                                case 45: $back = 'magenta'; break;
-                                                case 46: $back = 'cyan'; break;
-                                                case 47: $back = 'white'; break;
+                                                case '40': $back = 'black'; break;
+                                                case '41': $back = 'red'; break;
+                                                case '42': $back = 'green'; break;
+                                                case '43': $back = 'yellow'; break;
+                                                case '44': $back = 'blue'; break;
+                                                case '45': $back = 'magenta'; break;
+                                                case '46': $back = 'cyan'; break;
+                                                case '47': $back = 'white'; break;
                                                 // @codingStandardsIgnoreEnd
 
                                                 default:
@@ -376,13 +352,13 @@ class ANSI
                 continue;
             }
 
-            $this->tokenization[count($this->tokenization) - 1].= $source[$i];
+            $this->tokenization[count($this->tokenization) - 1] .= $source[$i];
             switch ($source[$i]) {
                 case "\r":
                     $this->x = 0;
                     break;
                 case "\n":
-                    $this->_newLine();
+                    $this->newLine();
                     break;
                 case "\x08": // backspace
                     if ($this->x) {
@@ -403,7 +379,7 @@ class ANSI
                     //if (!strlen($this->tokenization[count($this->tokenization) - 1])) {
                     //    array_pop($this->tokenization);
                     //}
-                    $this->ansi.= "\x1B";
+                    $this->ansi .= "\x1B";
                     break;
                 default:
                     $this->attrs[$this->y][$this->x] = clone $this->attr_cell;
@@ -419,7 +395,7 @@ class ANSI
 
                     if ($this->x > $this->max_x) {
                         $this->x = 0;
-                        $this->_newLine();
+                        $this->newLine();
                     } else {
                         $this->x++;
                     }
@@ -432,19 +408,18 @@ class ANSI
      *
      * Also update the $this->screen and $this->history buffers
      *
-     * @access private
      */
-    function _newLine()
+    private function newLine()
     {
         //if ($this->y < $this->max_y) {
         //    $this->y++;
         //}
 
         while ($this->y >= $this->max_y) {
-            $this->history = array_merge($this->history, array(array_shift($this->screen)));
+            $this->history = array_merge($this->history, [array_shift($this->screen)]);
             $this->screen[] = '';
 
-            $this->history_attrs = array_merge($this->history_attrs, array(array_shift($this->attrs)));
+            $this->history_attrs = array_merge($this->history_attrs, [array_shift($this->attrs)]);
             $this->attrs[] = $this->attr_row;
 
             if (count($this->history) >= $this->max_history) {
@@ -460,10 +435,12 @@ class ANSI
     /**
      * Returns the current coordinate without preformating
      *
-     * @access private
+     * @param \stdClass $last_attr
+     * @param \stdClass $cur_attr
+     * @param string $char
      * @return string
      */
-    function _processCoordinate($last_attr, $cur_attr, $char)
+    private function processCoordinate(\stdClass $last_attr, \stdClass $cur_attr, $char)
     {
         $output = '';
 
@@ -471,7 +448,7 @@ class ANSI
             $close = $open = '';
             if ($last_attr->foreground != $cur_attr->foreground) {
                 if ($cur_attr->foreground != 'white') {
-                    $open.= '<span style="color: ' . $cur_attr->foreground . '">';
+                    $open .= '<span style="color: ' . $cur_attr->foreground . '">';
                 }
                 if ($last_attr->foreground != 'white') {
                     $close = '</span>' . $close;
@@ -479,7 +456,7 @@ class ANSI
             }
             if ($last_attr->background != $cur_attr->background) {
                 if ($cur_attr->background != 'black') {
-                    $open.= '<span style="background: ' . $cur_attr->background . '">';
+                    $open .= '<span style="background: ' . $cur_attr->background . '">';
                 }
                 if ($last_attr->background != 'black') {
                     $close = '</span>' . $close;
@@ -487,29 +464,29 @@ class ANSI
             }
             if ($last_attr->bold != $cur_attr->bold) {
                 if ($cur_attr->bold) {
-                    $open.= '<b>';
+                    $open .= '<b>';
                 } else {
                     $close = '</b>' . $close;
                 }
             }
             if ($last_attr->underline != $cur_attr->underline) {
                 if ($cur_attr->underline) {
-                    $open.= '<u>';
+                    $open .= '<u>';
                 } else {
                     $close = '</u>' . $close;
                 }
             }
             if ($last_attr->blink != $cur_attr->blink) {
                 if ($cur_attr->blink) {
-                    $open.= '<blink>';
+                    $open .= '<blink>';
                 } else {
                     $close = '</blink>' . $close;
                 }
             }
-            $output.= $close . $open;
+            $output .= $close . $open;
         }
 
-        $output.= htmlspecialchars($char);
+        $output .= htmlspecialchars($char);
 
         return $output;
     }
@@ -517,59 +494,56 @@ class ANSI
     /**
      * Returns the current screen without preformating
      *
-     * @access private
      * @return string
      */
-    function _getScreen()
+    private function getScreenHelper()
     {
         $output = '';
         $last_attr = $this->base_attr_cell;
         for ($i = 0; $i <= $this->max_y; $i++) {
             for ($j = 0; $j <= $this->max_x; $j++) {
                 $cur_attr = $this->attrs[$i][$j];
-                $output.= $this->_processCoordinate($last_attr, $cur_attr, isset($this->screen[$i][$j]) ? $this->screen[$i][$j] : '');
+                $output .= $this->processCoordinate($last_attr, $cur_attr, isset($this->screen[$i][$j]) ? $this->screen[$i][$j] : '');
                 $last_attr = $this->attrs[$i][$j];
             }
-            $output.= "\r\n";
+            $output .= "\r\n";
         }
         $output = substr($output, 0, -2);
         // close any remaining open tags
-        $output.= $this->_processCoordinate($last_attr, $this->base_attr_cell, '');
+        $output .= $this->processCoordinate($last_attr, $this->base_attr_cell, '');
         return rtrim($output);
     }
 
     /**
      * Returns the current screen
      *
-     * @access public
      * @return string
      */
-    function getScreen()
+    public function getScreen()
     {
-        return '<pre width="' . ($this->max_x + 1) . '" style="color: white; background: black">' . $this->_getScreen() . '</pre>';
+        return '<pre width="' . ($this->max_x + 1) . '" style="color: white; background: black">' . $this->getScreenHelper() . '</pre>';
     }
 
     /**
      * Returns the current screen and the x previous lines
      *
-     * @access public
      * @return string
      */
-    function getHistory()
+    public function getHistory()
     {
         $scrollback = '';
         $last_attr = $this->base_attr_cell;
         for ($i = 0; $i < count($this->history); $i++) {
             for ($j = 0; $j <= $this->max_x + 1; $j++) {
                 $cur_attr = $this->history_attrs[$i][$j];
-                $scrollback.= $this->_processCoordinate($last_attr, $cur_attr, isset($this->history[$i][$j]) ? $this->history[$i][$j] : '');
+                $scrollback .= $this->processCoordinate($last_attr, $cur_attr, isset($this->history[$i][$j]) ? $this->history[$i][$j] : '');
                 $last_attr = $this->history_attrs[$i][$j];
             }
-            $scrollback.= "\r\n";
+            $scrollback .= "\r\n";
         }
         $base_attr_cell = $this->base_attr_cell;
         $this->base_attr_cell = $last_attr;
-        $scrollback.= $this->_getScreen();
+        $scrollback .= $this->getScreen();
         $this->base_attr_cell = $base_attr_cell;
 
         return '<pre width="' . ($this->max_x + 1) . '" style="color: white; background: black">' . $scrollback . '</span></pre>';
