@@ -9,52 +9,38 @@
  * utilized scheme is DER or the "Distinguished Encoding Rules".  PEM's are base64 encoded
  * DER blobs.
  *
- * \phpseclib\File\ASN1 decodes and encodes DER formatted messages and places them in a semantic context.
+ * \phpseclib3\File\ASN1 decodes and encodes DER formatted messages and places them in a semantic context.
  *
  * Uses the 1988 ASN.1 syntax.
  *
- * @category  File
- * @package   ASN1
  * @author    Jim Wigginton <terrafrost@php.net>
  * @copyright 2012 Jim Wigginton
  * @license   http://www.opensource.org/licenses/mit-license.html  MIT License
  * @link      http://phpseclib.sourceforge.net
  */
 
-namespace phpseclib\File;
+namespace phpseclib3\File;
 
-use phpseclib\File\ASN1\Element;
-use phpseclib\Math\BigInteger;
-use DateTime;
-use DateTimeZone;
+use phpseclib3\Common\Functions\Strings;
+use phpseclib3\File\ASN1\Element;
+use phpseclib3\Math\BigInteger;
 
 /**
  * Pure-PHP ASN.1 Parser
  *
- * @package ASN1
  * @author  Jim Wigginton <terrafrost@php.net>
- * @access  public
  */
-class ASN1
+abstract class ASN1
 {
-    /**#@+
-     * Tag Classes
-     *
-     * @access private
-     * @link http://www.itu.int/ITU-T/studygroups/com17/languages/X.690-0207.pdf#page=12
-     */
+    // Tag Classes
+    // http://www.itu.int/ITU-T/studygroups/com17/languages/X.690-0207.pdf#page=12
     const CLASS_UNIVERSAL        = 0;
     const CLASS_APPLICATION      = 1;
     const CLASS_CONTEXT_SPECIFIC = 2;
     const CLASS_PRIVATE          = 3;
-    /**#@-*/
 
-    /**#@+
-     * Tag Classes
-     *
-     * @access private
-     * @link http://www.obj-sys.com/asn1tutorial/node124.html
-    */
+    // Tag Classes
+    // http://www.obj-sys.com/asn1tutorial/node124.html
     const TYPE_BOOLEAN           = 1;
     const TYPE_INTEGER           = 2;
     const TYPE_BIT_STRING        = 3;
@@ -70,13 +56,9 @@ class ASN1
     //const TYPE_RELATIVE_OID      = 13;
     const TYPE_SEQUENCE          = 16; // SEQUENCE OF
     const TYPE_SET               = 17; // SET OF
-    /**#@-*/
-    /**#@+
-     * More Tag Classes
-     *
-     * @access private
-     * @link http://www.obj-sys.com/asn1tutorial/node10.html
-    */
+
+    // More Tag Classes
+    // http://www.obj-sys.com/asn1tutorial/node10.html
     const TYPE_NUMERIC_STRING   = 18;
     const TYPE_PRINTABLE_STRING = 19;
     const TYPE_TELETEX_STRING   = 20; // T61String
@@ -90,47 +72,34 @@ class ASN1
     const TYPE_UNIVERSAL_STRING = 28;
     //const TYPE_CHARACTER_STRING = 29;
     const TYPE_BMP_STRING       = 30;
-    /**#@-*/
 
-    /**#@+
-     * Tag Aliases
-     *
-     * These tags are kinda place holders for other tags.
-     *
-     * @access private
-    */
+    // Tag Aliases
+    // These tags are kinda place holders for other tags.
     const TYPE_CHOICE = -1;
     const TYPE_ANY    = -2;
-    /**#@-*/
 
     /**
-     * ASN.1 object identifier
+     * ASN.1 object identifiers
      *
      * @var array
-     * @access private
      * @link http://en.wikipedia.org/wiki/Object_identifier
      */
-    var $oids = array();
+    private static $oids = [];
+
+    /**
+     * ASN.1 object identifier reverse mapping
+     *
+     * @var array
+     */
+    private static $reverseOIDs = [];
 
     /**
      * Default date format
      *
      * @var string
-     * @access private
      * @link http://php.net/class.datetime
      */
-    var $format = 'D, d M Y H:i:s O';
-
-    /**
-     * Default date format
-     *
-     * @var array
-     * @access private
-     * @see self::setTimeFormat()
-     * @see self::asn1map()
-     * @link http://php.net/class.datetime
-     */
-    var $encoded;
+    private static $format = 'D, d M Y H:i:s O';
 
     /**
      * Filters
@@ -138,22 +107,40 @@ class ASN1
      * If the mapping type is self::TYPE_ANY what do we actually encode it as?
      *
      * @var array
-     * @access private
-     * @see self::_encode_der()
+     * @see self::encode_der()
      */
-    var $filters;
+    private static $filters;
+
+    /**
+     * Current Location of most recent ASN.1 encode process
+     *
+     * Useful for debug purposes
+     *
+     * @var array
+     * @see self::encode_der()
+     */
+    private static $location;
+
+    /**
+     * DER Encoded String
+     *
+     * In case we need to create ASN1\Element object's..
+     *
+     * @var string
+     * @see self::decodeDER()
+     */
+    private static $encoded;
 
     /**
      * Type mapping table for the ANY type.
      *
-     * Structured or unknown types are mapped to a \phpseclib\File\ASN1\Element.
+     * Structured or unknown types are mapped to a \phpseclib3\File\ASN1\Element.
      * Unambiguous types get the direct mapping (int/real/bool).
      * Others are mapped as a choice, with an extra indexing level.
      *
      * @var array
-     * @access public
      */
-    var $ANYmap = array(
+    const ANY_MAP = [
         self::TYPE_BOOLEAN              => true,
         self::TYPE_INTEGER              => true,
         self::TYPE_BIT_STRING           => 'bitString',
@@ -176,7 +163,7 @@ class ASN1
         self::TYPE_UNIVERSAL_STRING     => 'universalString',
         //self::TYPE_CHARACTER_STRING     => 'characterString',
         self::TYPE_BMP_STRING           => 'bmpString'
-    );
+    ];
 
     /**
      * String type to character size mapping table.
@@ -185,9 +172,8 @@ class ASN1
      * size == 0 indicates variable length encoding.
      *
      * @var array
-     * @access public
      */
-    var $stringTypeSize = array(
+    const STRING_TYPE_SIZE = [
         self::TYPE_UTF8_STRING      => 0,
         self::TYPE_BMP_STRING       => 2,
         self::TYPE_UNIVERSAL_STRING => 4,
@@ -195,26 +181,30 @@ class ASN1
         self::TYPE_TELETEX_STRING   => 1,
         self::TYPE_IA5_STRING       => 1,
         self::TYPE_VISIBLE_STRING   => 1,
-    );
+    ];
 
     /**
      * Parse BER-encoding
      *
      * Serves a similar purpose to openssl's asn1parse
      *
-     * @param string $encoded
-     * @return array
-     * @access public
+     * @param Element|string $encoded
+     * @return ?array
      */
-    function decodeBER($encoded)
+    public static function decodeBER($encoded)
     {
         if ($encoded instanceof Element) {
             $encoded = $encoded->element;
         }
 
-        $this->encoded = $encoded;
-        // encapsulate in an array for BC with the old decodeBER
-        return array($this->_decode_ber($encoded));
+        self::$encoded = $encoded;
+
+        $decoded = self::decode_ber($encoded);
+        if ($decoded === false) {
+            return null;
+        }
+
+        return [$decoded];
     }
 
     /**
@@ -227,15 +217,17 @@ class ASN1
      * @param string $encoded
      * @param int $start
      * @param int $encoded_pos
-     * @return array
-     * @access private
+     * @return array|bool
      */
-    function _decode_ber($encoded, $start = 0, $encoded_pos = 0)
+    private static function decode_ber($encoded, $start = 0, $encoded_pos = 0)
     {
-        $current = array('start' => $start);
+        $current = ['start' => $start];
 
+        if (!isset($encoded[$encoded_pos])) {
+            return false;
+        }
         $type = ord($encoded[$encoded_pos++]);
-        $start++;
+        $startOffset = 1;
 
         $constructed = ($type >> 5) & 1;
 
@@ -244,15 +236,28 @@ class ASN1
             $tag = 0;
             // process septets (since the eighth bit is ignored, it's not an octet)
             do {
+                if (!isset($encoded[$encoded_pos])) {
+                    return false;
+                }
                 $temp = ord($encoded[$encoded_pos++]);
+                $startOffset++;
                 $loop = $temp >> 7;
                 $tag <<= 7;
-                $tag |= $temp & 0x7F;
-                $start++;
+                $temp &= 0x7F;
+                // "bits 7 to 1 of the first subsequent octet shall not all be zero"
+                if ($startOffset == 2 && $temp == 0) {
+                    return false;
+                }
+                $tag |= $temp;
             } while ($loop);
         }
 
+        $start += $startOffset;
+
         // Length, as discussed in paragraph 8.1.3 of X.690-0207.pdf#page=13
+        if (!isset($encoded[$encoded_pos])) {
+            return false;
+        }
         $length = ord($encoded[$encoded_pos++]);
         $start++;
         if ($length == 0x80) { // indefinite length
@@ -262,15 +267,16 @@ class ASN1
         } elseif ($length & 0x80) { // definite length, long form
             // technically, the long form of the length can be represented by up to 126 octets (bytes), but we'll only
             // support it up to four.
-            $length&= 0x7F;
+            $length &= 0x7F;
             $temp = substr($encoded, $encoded_pos, $length);
             $encoded_pos += $length;
             // tags of indefinte length don't really have a header length; this length includes the tag
-            $current+= array('headerlength' => $length + 2);
-            $start+= $length;
+            $current += ['headerlength' => $length + 2];
+            $start += $length;
             extract(unpack('Nlength', substr(str_pad($temp, 4, chr(0), STR_PAD_LEFT), -4)));
+            /** @var integer $length */
         } else {
-            $current+= array('headerlength' => 2);
+            $current += ['headerlength' => 2];
         }
 
         if ($length > (strlen($encoded) - $encoded_pos)) {
@@ -297,36 +303,36 @@ class ASN1
             case self::CLASS_PRIVATE:
             case self::CLASS_CONTEXT_SPECIFIC:
                 if (!$constructed) {
-                    return array(
+                    return [
                         'type'     => $class,
                         'constant' => $tag,
                         'content'  => $content,
                         'length'   => $length + $start - $current['start']
-                    );
+                    ] + $current;
                 }
 
-                $newcontent = array();
+                $newcontent = [];
                 $remainingLength = $length;
                 while ($remainingLength > 0) {
-                    $temp = $this->_decode_ber($content, $start, $content_pos);
+                    $temp = self::decode_ber($content, $start, $content_pos);
                     if ($temp === false) {
                         break;
                     }
                     $length = $temp['length'];
                     // end-of-content octets - see paragraph 8.1.5
                     if (substr($content, $content_pos + $length, 2) == "\0\0") {
-                        $length+= 2;
-                        $start+= $length;
+                        $length += 2;
+                        $start += $length;
                         $newcontent[] = $temp;
                         break;
                     }
-                    $start+= $length;
-                    $remainingLength-= $length;
+                    $start += $length;
+                    $remainingLength -= $length;
                     $newcontent[] = $temp;
                     $content_pos += $length;
                 }
 
-                return array(
+                return [
                     'type'     => $class,
                     'constant' => $tag,
                     // the array encapsulation is for BC with the old format
@@ -335,22 +341,25 @@ class ASN1
                     // the absence of $content['headerlength'] is how we know if something is indefinite or not.
                     // technically, it could be defined to be 2 and then another indicator could be used but whatever.
                     'length'   => $start - $current['start']
-                ) + $current;
+                ] + $current;
         }
 
-        $current+= array('type' => $tag);
+        $current += ['type' => $tag];
 
         // decode UNIVERSAL tags
         switch ($tag) {
             case self::TYPE_BOOLEAN:
                 // "The contents octets shall consist of a single octet." -- paragraph 8.2.1
-                //if (strlen($content) != 1) {
-                //    return false;
-                //}
+                if ($constructed || strlen($content) != 1) {
+                    return false;
+                }
                 $current['content'] = (bool) ord($content[$content_pos]);
                 break;
             case self::TYPE_INTEGER:
             case self::TYPE_ENUMERATED:
+                if ($constructed) {
+                    return false;
+                }
                 $current['content'] = new BigInteger(substr($content, $content_pos), -256);
                 break;
             case self::TYPE_REAL: // not currently supported
@@ -362,23 +371,23 @@ class ASN1
                 if (!$constructed) {
                     $current['content'] = substr($content, $content_pos);
                 } else {
-                    $temp = $this->_decode_ber($content, $start, $content_pos);
+                    $temp = self::decode_ber($content, $start, $content_pos);
                     if ($temp === false) {
                         return false;
                     }
-                    $length-= (strlen($content) - $content_pos);
+                    $length -= (strlen($content) - $content_pos);
                     $last = count($temp) - 1;
                     for ($i = 0; $i < $last; $i++) {
                         // all subtags should be bit strings
-                        //if ($temp[$i]['type'] != self::TYPE_BIT_STRING) {
-                        //    return false;
-                        //}
-                        $current['content'].= substr($temp[$i]['content'], 1);
+                        if ($temp[$i]['type'] != self::TYPE_BIT_STRING) {
+                            return false;
+                        }
+                        $current['content'] .= substr($temp[$i]['content'], 1);
                     }
                     // all subtags should be bit strings
-                    //if ($temp[$last]['type'] != self::TYPE_BIT_STRING) {
-                    //    return false;
-                    //}
+                    if ($temp[$last]['type'] != self::TYPE_BIT_STRING) {
+                        return false;
+                    }
                     $current['content'] = $temp[$last]['content'][0] . $current['content'] . substr($temp[$i]['content'], 1);
                 }
                 break;
@@ -389,33 +398,36 @@ class ASN1
                     $current['content'] = '';
                     $length = 0;
                     while (substr($content, $content_pos, 2) != "\0\0") {
-                        $temp = $this->_decode_ber($content, $length + $start, $content_pos);
+                        $temp = self::decode_ber($content, $length + $start, $content_pos);
                         if ($temp === false) {
                             return false;
                         }
                         $content_pos += $temp['length'];
                         // all subtags should be octet strings
-                        //if ($temp['type'] != self::TYPE_OCTET_STRING) {
-                        //    return false;
-                        //}
-                        $current['content'].= $temp['content'];
-                        $length+= $temp['length'];
+                        if ($temp['type'] != self::TYPE_OCTET_STRING) {
+                            return false;
+                        }
+                        $current['content'] .= $temp['content'];
+                        $length += $temp['length'];
                     }
                     if (substr($content, $content_pos, 2) == "\0\0") {
-                        $length+= 2; // +2 for the EOC
+                        $length += 2; // +2 for the EOC
                     }
                 }
                 break;
             case self::TYPE_NULL:
                 // "The contents octets shall not contain any octets." -- paragraph 8.8.2
-                //if (strlen($content)) {
-                //    return false;
-                //}
+                if ($constructed || strlen($content)) {
+                    return false;
+                }
                 break;
             case self::TYPE_SEQUENCE:
             case self::TYPE_SET:
+                if (!$constructed) {
+                    return false;
+                }
                 $offset = 0;
-                $current['content'] = array();
+                $current['content'] = [];
                 $content_len = strlen($content);
                 while ($content_pos < $content_len) {
                     // if indefinite length construction was used and we have an end-of-content string next
@@ -424,17 +436,23 @@ class ASN1
                         $length = $offset + 2; // +2 for the EOC
                         break 2;
                     }
-                    $temp = $this->_decode_ber($content, $start + $offset, $content_pos);
+                    $temp = self::decode_ber($content, $start + $offset, $content_pos);
                     if ($temp === false) {
                         return false;
                     }
                     $content_pos += $temp['length'];
                     $current['content'][] = $temp;
-                    $offset+= $temp['length'];
+                    $offset += $temp['length'];
                 }
                 break;
             case self::TYPE_OBJECT_IDENTIFIER:
-                $current['content'] = $this->_decodeOID(substr($content, $content_pos));
+                if ($constructed) {
+                    return false;
+                }
+                $current['content'] = self::decodeOID(substr($content, $content_pos));
+                if ($current['content'] === false) {
+                    return false;
+                }
                 break;
             /* Each character string type shall be encoded as if it had been declared:
                [UNIVERSAL x] IMPLICIT OCTET STRING
@@ -464,18 +482,26 @@ class ASN1
             case self::TYPE_UTF8_STRING:
                 // ????
             case self::TYPE_BMP_STRING:
+                if ($constructed) {
+                    return false;
+                }
                 $current['content'] = substr($content, $content_pos);
                 break;
             case self::TYPE_UTC_TIME:
             case self::TYPE_GENERALIZED_TIME:
-                $current['content'] = $this->_decodeTime(substr($content, $content_pos), $tag);
+                if ($constructed) {
+                    return false;
+                }
+                $current['content'] = self::decodeTime(substr($content, $content_pos), $tag);
+                break;
             default:
+                return false;
         }
 
-        $start+= $length;
+        $start += $length;
 
         // ie. length is the length of the full TLV encoding - it's not just the length of the value
-        return $current + array('length' => $start - $current['start']);
+        return $current + ['length' => $start - $current['start']];
     }
 
     /**
@@ -488,15 +514,10 @@ class ASN1
      * @param array $decoded
      * @param array $mapping
      * @param array $special
-     * @return array
-     * @access public
+     * @return array|bool|Element|string|null
      */
-    function asn1map($decoded, $mapping, $special = array())
+    public static function asn1map(array $decoded, $mapping, $special = [])
     {
-        if (!is_array($decoded)) {
-            return false;
-        }
-
         if (isset($mapping['explicit']) && is_array($decoded['content'])) {
             $decoded = $decoded['content'][0];
         }
@@ -504,12 +525,13 @@ class ASN1
         switch (true) {
             case $mapping['type'] == self::TYPE_ANY:
                 $intype = $decoded['type'];
-                if (isset($decoded['constant']) || !isset($this->ANYmap[$intype]) || (ord($this->encoded[$decoded['start']]) & 0x20)) {
-                    return new Element(substr($this->encoded, $decoded['start'], $decoded['length']));
+                // !isset(self::ANY_MAP[$intype]) produces a fatal error on PHP 5.6
+                if (isset($decoded['constant']) || !array_key_exists($intype, self::ANY_MAP) || (ord(self::$encoded[$decoded['start']]) & 0x20)) {
+                    return new Element(substr(self::$encoded, $decoded['start'], $decoded['length']));
                 }
-                $inmap = $this->ANYmap[$intype];
+                $inmap = self::ANY_MAP[$intype];
                 if (is_string($inmap)) {
-                    return array($inmap => $this->asn1map($decoded, array('type' => $intype) + $mapping, $special));
+                    return [$inmap => self::asn1map($decoded, ['type' => $intype] + $mapping, $special)];
                 }
                 break;
             case $mapping['type'] == self::TYPE_CHOICE:
@@ -517,19 +539,19 @@ class ASN1
                     switch (true) {
                         case isset($option['constant']) && $option['constant'] == $decoded['constant']:
                         case !isset($option['constant']) && $option['type'] == $decoded['type']:
-                            $value = $this->asn1map($decoded, $option, $special);
+                            $value = self::asn1map($decoded, $option, $special);
                             break;
                         case !isset($option['constant']) && $option['type'] == self::TYPE_CHOICE:
-                            $v = $this->asn1map($decoded, $option, $special);
+                            $v = self::asn1map($decoded, $option, $special);
                             if (isset($v)) {
                                 $value = $v;
                             }
                     }
                     if (isset($value)) {
                         if (isset($special[$key])) {
-                            $value = call_user_func($special[$key], $value);
+                            $value = $special[$key]($value);
                         }
-                        return array($key => $value);
+                        return [$key => $value];
                     }
                 }
                 return null;
@@ -555,13 +577,13 @@ class ASN1
 
         switch ($decoded['type']) {
             case self::TYPE_SEQUENCE:
-                $map = array();
+                $map = [];
 
                 // ignore the min and max
                 if (isset($mapping['min']) && isset($mapping['max'])) {
                     $child = $mapping['children'];
                     foreach ($decoded['content'] as $content) {
-                        if (($map[] = $this->asn1map($content, $child, $special)) === null) {
+                        if (($map[] = self::asn1map($content, $child, $special)) === null) {
                             return null;
                         }
                     }
@@ -597,43 +619,43 @@ class ASN1
                                 $maymatch = $constant == $temp['constant'] && $childClass == $tempClass;
                             } else {
                                 // Can only match if no constant expected and type matches or is generic.
-                                $maymatch = !isset($child['constant']) && array_search($child['type'], array($temp['type'], self::TYPE_ANY, self::TYPE_CHOICE)) !== false;
+                                $maymatch = !isset($child['constant']) && array_search($child['type'], [$temp['type'], self::TYPE_ANY, self::TYPE_CHOICE]) !== false;
                             }
                         }
                     }
 
                     if ($maymatch) {
                         // Attempt submapping.
-                        $candidate = $this->asn1map($temp, $child, $special);
+                        $candidate = self::asn1map($temp, $child, $special);
                         $maymatch = $candidate !== null;
                     }
 
                     if ($maymatch) {
                         // Got the match: use it.
                         if (isset($special[$key])) {
-                            $candidate = call_user_func($special[$key], $candidate);
+                            $candidate = $special[$key]($candidate);
                         }
                         $map[$key] = $candidate;
                         $i++;
                     } elseif (isset($child['default'])) {
-                        $map[$key] = $child['default']; // Use default.
+                        $map[$key] = $child['default'];
                     } elseif (!isset($child['optional'])) {
                         return null; // Syntax error.
                     }
                 }
 
                 // Fail mapping if all input items have not been consumed.
-                return $i < $n ? null: $map;
+                return $i < $n ? null : $map;
 
             // the main diff between sets and sequences is the encapsulation of the foreach in another for loop
             case self::TYPE_SET:
-                $map = array();
+                $map = [];
 
                 // ignore the min and max
                 if (isset($mapping['min']) && isset($mapping['max'])) {
                     $child = $mapping['children'];
                     foreach ($decoded['content'] as $content) {
-                        if (($map[] = $this->asn1map($content, $child, $special)) === null) {
+                        if (($map[] = self::asn1map($content, $child, $special)) === null) {
                             return null;
                         }
                     }
@@ -669,13 +691,13 @@ class ASN1
                                 $maymatch = $constant == $temp['constant'] && $childClass == $tempClass;
                             } else {
                                 // Can only match if no constant expected and type matches or is generic.
-                                $maymatch = !isset($child['constant']) && array_search($child['type'], array($temp['type'], self::TYPE_ANY, self::TYPE_CHOICE)) !== false;
+                                $maymatch = !isset($child['constant']) && array_search($child['type'], [$temp['type'], self::TYPE_ANY, self::TYPE_CHOICE]) !== false;
                             }
                         }
 
                         if ($maymatch) {
                             // Attempt submapping.
-                            $candidate = $this->asn1map($temp, $child, $special);
+                            $candidate = self::asn1map($temp, $child, $special);
                             $maymatch = $candidate !== null;
                         }
 
@@ -685,7 +707,7 @@ class ASN1
 
                         // Got the match: use it.
                         if (isset($special[$key])) {
-                            $candidate = call_user_func($special[$key], $candidate);
+                            $candidate = $special[$key]($candidate);
                         }
                         $map[$key] = $candidate;
                         break;
@@ -703,7 +725,7 @@ class ASN1
                 }
                 return $map;
             case self::TYPE_OBJECT_IDENTIFIER:
-                return isset($this->oids[$decoded['content']]) ? $this->oids[$decoded['content']] : $decoded['content'];
+                return isset(self::$oids[$decoded['content']]) ? self::$oids[$decoded['content']] : $decoded['content'];
             case self::TYPE_UTC_TIME:
             case self::TYPE_GENERALIZED_TIME:
                 // for explicitly tagged optional stuff
@@ -714,9 +736,9 @@ class ASN1
                 // in theory, doing isset($mapping['implicit']) would work but malformed certs do exist
                 // in the wild that OpenSSL decodes without issue so we'll support them as well
                 if (!is_object($decoded['content'])) {
-                    $decoded['content'] = $this->_decodeTime($decoded['content'], $decoded['type']);
+                    $decoded['content'] = self::decodeTime($decoded['content'], $decoded['type']);
                 }
-                return $decoded['content'] ? $decoded['content']->format($this->format) : false;
+                return $decoded['content'] ? $decoded['content']->format(self::$format) : false;
             case self::TYPE_BIT_STRING:
                 if (isset($mapping['mapping'])) {
                     $offset = ord($decoded['content'][0]);
@@ -729,7 +751,7 @@ class ASN1
                         therefore ensure that different semantics are not associated with such values which differ only in the number of trailing
                         0 bits."
                     */
-                    $bits = count($mapping['mapping']) == $size ? array() : array_fill(0, count($mapping['mapping']) - $size, false);
+                    $bits = count($mapping['mapping']) == $size ? [] : array_fill(0, count($mapping['mapping']) - $size, false);
                     for ($i = strlen($decoded['content']) - 1; $i > 0; $i--) {
                         $current = ord($decoded['content'][$i]);
                         for ($j = $offset; $j < 8; $j++) {
@@ -737,7 +759,7 @@ class ASN1
                         }
                         $offset = 0;
                     }
-                    $values = array();
+                    $values = [];
                     $map = array_reverse($mapping['mapping']);
                     foreach ($map as $i => $value) {
                         if ($bits[$i]) {
@@ -746,12 +768,12 @@ class ASN1
                     }
                     return $values;
                 }
+                // fall-through
             case self::TYPE_OCTET_STRING:
-                return base64_encode($decoded['content']);
+                return $decoded['content'];
             case self::TYPE_NULL:
                 return '';
             case self::TYPE_BOOLEAN:
-                return $decoded['content'];
             case self::TYPE_NUMERIC_STRING:
             case self::TYPE_PRINTABLE_STRING:
             case self::TYPE_TELETEX_STRING:
@@ -781,6 +803,26 @@ class ASN1
     }
 
     /**
+     * DER-decode the length
+     *
+     * DER supports lengths up to (2**8)**127, however, we'll only support lengths up to (2**8)**4.  See
+     * {@link http://itu.int/ITU-T/studygroups/com17/languages/X.690-0207.pdf#p=13 X.690 paragraph 8.1.3} for more information.
+     *
+     * @param string $string
+     * @return int
+     */
+    public static function decodeLength(&$string)
+    {
+        $length = ord(Strings::shift($string));
+        if ($length & 0x80) { // definite length, long form
+            $length &= 0x7F;
+            $temp = Strings::shift($string, $length);
+            list(, $length) = unpack('N', substr(str_pad($temp, 4, chr(0), STR_PAD_LEFT), -4));
+        }
+        return $length;
+    }
+
+    /**
      * ASN.1 Encode
      *
      * DER-encodes an ASN.1 semantic mapping ($mapping).  Some libraries would probably call this function
@@ -788,28 +830,27 @@ class ASN1
      *
      * "Special" mappings can be applied via $special.
      *
-     * @param string $source
-     * @param string $mapping
-     * @param int $idx
+     * @param Element|string|array $source
+     * @param array $mapping
+     * @param array $special
      * @return string
-     * @access public
      */
-    function encodeDER($source, $mapping, $special = array())
+    public static function encodeDER($source, $mapping, $special = [])
     {
-        $this->location = array();
-        return $this->_encode_der($source, $mapping, null, $special);
+        self::$location = [];
+        return self::encode_der($source, $mapping, null, $special);
     }
 
     /**
      * ASN.1 Encode (Helper function)
      *
-     * @param string $source
-     * @param string $mapping
+     * @param Element|string|array|null $source
+     * @param array $mapping
      * @param int $idx
+     * @param array $special
      * @return string
-     * @access private
      */
-    function _encode_der($source, $mapping, $idx = null, $special = array())
+    private static function encode_der($source, array $mapping, $idx = null, array $special = [])
     {
         if ($source instanceof Element) {
             return $source->element;
@@ -822,9 +863,9 @@ class ASN1
 
         if (isset($idx)) {
             if (isset($special[$idx])) {
-                $source = call_user_func($special[$idx], $source);
+                $source = $special[$idx]($source);
             }
-            $this->location[] = $idx;
+            self::$location[] = $idx;
         }
 
         $tag = $mapping['type'];
@@ -832,19 +873,19 @@ class ASN1
         switch ($tag) {
             case self::TYPE_SET:    // Children order is not important, thus process in sequence.
             case self::TYPE_SEQUENCE:
-                $tag|= 0x20; // set the constructed bit
+                $tag |= 0x20; // set the constructed bit
 
                 // ignore the min and max
                 if (isset($mapping['min']) && isset($mapping['max'])) {
-                    $value = array();
+                    $value = [];
                     $child = $mapping['children'];
 
                     foreach ($source as $content) {
-                        $temp = $this->_encode_der($content, $child, null, $special);
+                        $temp = self::encode_der($content, $child, null, $special);
                         if ($temp === false) {
                             return false;
                         }
-                        $value[]= $temp;
+                        $value[] = $temp;
                     }
                     /* "The encodings of the component values of a set-of value shall appear in ascending order, the encodings being compared
                         as octet strings with the shorter components being padded at their trailing end with 0-octets.
@@ -867,7 +908,7 @@ class ASN1
                         continue;
                     }
 
-                    $temp = $this->_encode_der($source[$key], $child, $key, $special);
+                    $temp = self::encode_der($source[$key], $child, $key, $special);
                     if ($temp === false) {
                         return false;
                     }
@@ -891,13 +932,13 @@ class ASN1
                          */
                         if (isset($child['explicit']) || $child['type'] == self::TYPE_CHOICE) {
                             $subtag = chr((self::CLASS_CONTEXT_SPECIFIC << 6) | 0x20 | $child['constant']);
-                            $temp = $subtag . $this->_encodeLength(strlen($temp)) . $temp;
+                            $temp = $subtag . self::encodeLength(strlen($temp)) . $temp;
                         } else {
                             $subtag = chr((self::CLASS_CONTEXT_SPECIFIC << 6) | (ord($temp[0]) & 0x20) | $child['constant']);
                             $temp = $subtag . substr($temp, 1);
                         }
                     }
-                    $value.= $temp;
+                    $value .= $temp;
                 }
                 break;
             case self::TYPE_CHOICE:
@@ -908,7 +949,7 @@ class ASN1
                         continue;
                     }
 
-                    $temp = $this->_encode_der($source[$key], $child, $key, $special);
+                    $temp = self::encode_der($source[$key], $child, $key, $special);
                     if ($temp === false) {
                         return false;
                     }
@@ -925,7 +966,7 @@ class ASN1
                     if (isset($child['constant'])) {
                         if (isset($child['explicit']) || $child['type'] == self::TYPE_CHOICE) {
                             $subtag = chr((self::CLASS_CONTEXT_SPECIFIC << 6) | 0x20 | $child['constant']);
-                            $temp = $subtag . $this->_encodeLength(strlen($temp)) . $temp;
+                            $temp = $subtag . self::encodeLength(strlen($temp)) . $temp;
                         } else {
                             $subtag = chr((self::CLASS_CONTEXT_SPECIFIC << 6) | (ord($temp[0]) & 0x20) | $child['constant']);
                             $temp = $subtag . substr($temp, 1);
@@ -934,7 +975,7 @@ class ASN1
                 }
 
                 if (isset($idx)) {
-                    array_pop($this->location);
+                    array_pop(self::$location);
                 }
 
                 if ($temp && isset($mapping['cast'])) {
@@ -964,8 +1005,11 @@ class ASN1
             case self::TYPE_UTC_TIME:
             case self::TYPE_GENERALIZED_TIME:
                 $format = $mapping['type'] == self::TYPE_UTC_TIME ? 'y' : 'Y';
-                $format.= 'mdHis';
-                $date = new DateTime($source, new DateTimeZone('GMT'));
+                $format .= 'mdHis';
+                // if $source does _not_ include timezone information within it then assume that the timezone is GMT
+                $date = new \DateTime($source, new \DateTimeZone('GMT'));
+                // if $source _does_ include timezone information within it then convert the time to GMT
+                $date->setTimezone(new \DateTimeZone('GMT'));
                 $value = $date->format($format) . 'Z';
                 break;
             case self::TYPE_BIT_STRING:
@@ -995,46 +1039,47 @@ class ASN1
                     $bits = implode('', array_pad($bits, $size + $offset + 1, 0));
                     $bytes = explode(' ', rtrim(chunk_split($bits, 8, ' ')));
                     foreach ($bytes as $byte) {
-                        $value.= chr(bindec($byte));
+                        $value .= chr(bindec($byte));
                     }
 
                     break;
                 }
+                // fall-through
             case self::TYPE_OCTET_STRING:
                 /* The initial octet shall encode, as an unsigned binary integer with bit 1 as the least significant bit,
                    the number of unused bits in the final subsequent octet. The number shall be in the range zero to seven.
 
                    -- http://www.itu.int/ITU-T/studygroups/com17/languages/X.690-0207.pdf#page=16 */
-                $value = base64_decode($source);
+                $value = $source;
                 break;
             case self::TYPE_OBJECT_IDENTIFIER:
-                $value = $this->_encodeOID($source);
+                $value = self::encodeOID($source);
                 break;
             case self::TYPE_ANY:
-                $loc = $this->location;
+                $loc = self::$location;
                 if (isset($idx)) {
-                    array_pop($this->location);
+                    array_pop(self::$location);
                 }
 
                 switch (true) {
                     case !isset($source):
-                        return $this->_encode_der(null, array('type' => self::TYPE_NULL) + $mapping, null, $special);
+                        return self::encode_der(null, ['type' => self::TYPE_NULL] + $mapping, null, $special);
                     case is_int($source):
                     case $source instanceof BigInteger:
-                        return $this->_encode_der($source, array('type' => self::TYPE_INTEGER) + $mapping, null, $special);
+                        return self::encode_der($source, ['type' => self::TYPE_INTEGER] + $mapping, null, $special);
                     case is_float($source):
-                        return $this->_encode_der($source, array('type' => self::TYPE_REAL) + $mapping, null, $special);
+                        return self::encode_der($source, ['type' => self::TYPE_REAL] + $mapping, null, $special);
                     case is_bool($source):
-                        return $this->_encode_der($source, array('type' => self::TYPE_BOOLEAN) + $mapping, null, $special);
+                        return self::encode_der($source, ['type' => self::TYPE_BOOLEAN] + $mapping, null, $special);
                     case is_array($source) && count($source) == 1:
                         $typename = implode('', array_keys($source));
-                        $outtype = array_search($typename, $this->ANYmap, true);
+                        $outtype = array_search($typename, self::ANY_MAP, true);
                         if ($outtype !== false) {
-                            return $this->_encode_der($source[$typename], array('type' => $outtype) + $mapping, null, $special);
+                            return self::encode_der($source[$typename], ['type' => $outtype] + $mapping, null, $special);
                         }
                 }
 
-                $filters = $this->filters;
+                $filters = self::$filters;
                 foreach ($loc as $part) {
                     if (!isset($filters[$part])) {
                         $filters = false;
@@ -1043,10 +1088,9 @@ class ASN1
                     $filters = $filters[$part];
                 }
                 if ($filters === false) {
-                    user_error('No filters defined for ' . implode('/', $loc));
-                    return false;
+                    throw new \RuntimeException('No filters defined for ' . implode('/', $loc));
                 }
-                return $this->_encode_der($source, $filters + $mapping, null, $special);
+                return self::encode_der($source, $filters + $mapping, null, $special);
             case self::TYPE_NULL:
                 $value = '';
                 break;
@@ -1067,44 +1111,23 @@ class ASN1
                 $value = $source ? "\xFF" : "\x00";
                 break;
             default:
-                user_error('Mapping provides no type definition for ' . implode('/', $this->location));
-                return false;
+                throw new \RuntimeException('Mapping provides no type definition for ' . implode('/', self::$location));
         }
 
         if (isset($idx)) {
-            array_pop($this->location);
+            array_pop(self::$location);
         }
 
         if (isset($mapping['cast'])) {
             if (isset($mapping['explicit']) || $mapping['type'] == self::TYPE_CHOICE) {
-                $value = chr($tag) . $this->_encodeLength(strlen($value)) . $value;
+                $value = chr($tag) . self::encodeLength(strlen($value)) . $value;
                 $tag = ($mapping['class'] << 6) | 0x20 | $mapping['cast'];
             } else {
                 $tag = ($mapping['class'] << 6) | (ord($temp[0]) & 0x20) | $mapping['cast'];
             }
         }
 
-        return chr($tag) . $this->_encodeLength(strlen($value)) . $value;
-    }
-
-    /**
-     * DER-encode the length
-     *
-     * DER supports lengths up to (2**8)**127, however, we'll only support lengths up to (2**8)**4.  See
-     * {@link http://itu.int/ITU-T/studygroups/com17/languages/X.690-0207.pdf#p=13 X.690 paragraph 8.1.3} for more information.
-     *
-     * @access private
-     * @param int $length
-     * @return string
-     */
-    function _encodeLength($length)
-    {
-        if ($length <= 0x7F) {
-            return chr($length);
-        }
-
-        $temp = ltrim(pack('N', $length), chr(0));
-        return pack('Ca*', 0x80 | strlen($temp), $temp);
+        return chr($tag) . self::encodeLength(strlen($value)) . $value;
     }
 
     /**
@@ -1112,20 +1135,24 @@ class ASN1
      *
      * Called by _decode_ber()
      *
-     * @access private
      * @param string $content
      * @return string
      */
-    function _decodeOID($content)
+    public static function decodeOID($content)
     {
         static $eighty;
         if (!$eighty) {
             $eighty = new BigInteger(80);
         }
 
-        $oid = array();
+        $oid = [];
         $pos = 0;
         $len = strlen($content);
+
+        if (ord($content[$len - 1]) & 0x80) {
+            return false;
+        }
+
         $n = new BigInteger();
         while ($pos < $len) {
             $temp = ord($content[$pos++]);
@@ -1160,11 +1187,10 @@ class ASN1
      *
      * Called by _encode_der()
      *
-     * @access private
-     * @param string $content
+     * @param string $source
      * @return string
      */
-    function _encodeOID($source)
+    public static function encodeOID($source)
     {
         static $mask, $zero, $forty;
         if (!$mask) {
@@ -1173,11 +1199,15 @@ class ASN1
             $forty = new BigInteger(40);
         }
 
-        $oid = preg_match('#(?:\d+\.)+#', $source) ? $source : array_search($source, $this->oids);
-        if ($oid === false) {
-            user_error('Invalid OID');
-            return false;
+        if (!preg_match('#(?:\d+\.)+#', $source)) {
+            $oid = isset(self::$reverseOIDs[$source]) ? self::$reverseOIDs[$source] : false;
+        } else {
+            $oid = $source;
         }
+        if ($oid === false) {
+            throw new \RuntimeException('Invalid OID');
+        }
+
         $parts = explode('.', $oid);
         $part1 = array_shift($parts);
         $part2 = array_shift($parts);
@@ -1203,7 +1233,7 @@ class ASN1
                 }
                 $temp[strlen($temp) - 1] = $temp[strlen($temp) - 1] & chr(0x7F);
             }
-            $value.= $temp;
+            $value .= $temp;
         }
 
         return $value;
@@ -1214,12 +1244,11 @@ class ASN1
      *
      * Called by _decode_ber() and in the case of implicit tags asn1map().
      *
-     * @access private
      * @param string $content
      * @param int $tag
-     * @return string
+     * @return \DateTime|false
      */
-    function _decodeTime($content, $tag)
+    private static function decodeTime($content, $tag)
     {
         /* UTCTime:
            http://tools.ietf.org/html/rfc5280#section-4.1.2.5.1
@@ -1241,7 +1270,7 @@ class ASN1
             $prefix = substr($content, 0, 2) >= 50 ? '19' : '20';
             $content = $prefix . $content;
         } elseif (strpos($content, '.') !== false) {
-            $format.= '.u';
+            $format .= '.u';
         }
 
         if ($content[strlen($content) - 1] == 'Z') {
@@ -1249,12 +1278,12 @@ class ASN1
         }
 
         if (strpos($content, '-') !== false || strpos($content, '+') !== false) {
-            $format.= 'O';
+            $format .= 'O';
         }
 
         // error supression isn't necessary as of PHP 7.0:
         // http://php.net/manual/en/migration70.other-changes.php
-        return @DateTime::createFromFormat($format, $content);
+        return @\DateTime::createFromFormat($format, $content);
     }
 
     /**
@@ -1262,55 +1291,38 @@ class ASN1
      *
      * Sets the time / date format for asn1map().
      *
-     * @access public
      * @param string $format
      */
-    function setTimeFormat($format)
+    public static function setTimeFormat($format)
     {
-        $this->format = $format;
+        self::$format = $format;
     }
 
     /**
      * Load OIDs
      *
      * Load the relevant OIDs for a particular ASN.1 semantic mapping.
+     * Previously loaded OIDs are retained.
      *
-     * @access public
      * @param array $oids
      */
-    function loadOIDs($oids)
+    public static function loadOIDs(array $oids)
     {
-        $this->oids = $oids;
+        self::$reverseOIDs += $oids;
+        self::$oids = array_flip(self::$reverseOIDs);
     }
 
     /**
-     * Load filters
+     * Set filters
      *
-     * See \phpseclib\File\X509, etc, for an example.
+     * See \phpseclib3\File\X509, etc, for an example.
+     * Previously loaded filters are not retained.
      *
-     * @access public
      * @param array $filters
      */
-    function loadFilters($filters)
+    public static function setFilters(array $filters)
     {
-        $this->filters = $filters;
-    }
-
-    /**
-     * String Shift
-     *
-     * Inspired by array_shift
-     *
-     * @param string $string
-     * @param int $index
-     * @return string
-     * @access private
-     */
-    function _string_shift(&$string, $index = 1)
-    {
-        $substr = substr($string, 0, $index);
-        $string = substr($string, $index);
-        return $substr;
+        self::$filters = $filters;
     }
 
     /**
@@ -1323,15 +1335,15 @@ class ASN1
      * @param int $from
      * @param int $to
      * @return string
-     * @access public
      */
-    function convert($in, $from = self::TYPE_UTF8_STRING, $to = self::TYPE_UTF8_STRING)
+    public static function convert($in, $from = self::TYPE_UTF8_STRING, $to = self::TYPE_UTF8_STRING)
     {
-        if (!isset($this->stringTypeSize[$from]) || !isset($this->stringTypeSize[$to])) {
+        // isset(self::STRING_TYPE_SIZE[$from] returns a fatal error on PHP 5.6
+        if (!array_key_exists($from, self::STRING_TYPE_SIZE) || !array_key_exists($to, self::STRING_TYPE_SIZE)) {
             return false;
         }
-        $insize = $this->stringTypeSize[$from];
-        $outsize = $this->stringTypeSize[$to];
+        $insize = self::STRING_TYPE_SIZE[$from];
+        $outsize = self::STRING_TYPE_SIZE[$to];
         $inlength = strlen($in);
         $out = '';
 
@@ -1346,8 +1358,10 @@ class ASN1
                 case $insize == 4:
                     $c = ($c << 8) | ord($in[$i++]);
                     $c = ($c << 8) | ord($in[$i++]);
+                    // fall-through
                 case $insize == 2:
                     $c = ($c << 8) | ord($in[$i++]);
+                    // fall-through
                 case $insize == 1:
                     break;
                 case ($c & 0x80) == 0x00:
@@ -1376,9 +1390,11 @@ class ASN1
                     $c >>= 8;
                     $v .= chr($c & 0xFF);
                     $c >>= 8;
+                    // fall-through
                 case $outsize == 2:
                     $v .= chr($c & 0xFF);
                     $c >>= 8;
+                    // fall-through
                 case $outsize == 1:
                     $v .= chr($c & 0xFF);
                     $c >>= 8;
@@ -1391,18 +1407,23 @@ class ASN1
                 case $c >= 0x04000000:
                     $v .= chr(0x80 | ($c & 0x3F));
                     $c = ($c >> 6) | 0x04000000;
+                    // fall-through
                 case $c >= 0x00200000:
                     $v .= chr(0x80 | ($c & 0x3F));
                     $c = ($c >> 6) | 0x00200000;
+                    // fall-through
                 case $c >= 0x00010000:
                     $v .= chr(0x80 | ($c & 0x3F));
                     $c = ($c >> 6) | 0x00010000;
+                    // fall-through
                 case $c >= 0x00000800:
                     $v .= chr(0x80 | ($c & 0x3F));
                     $c = ($c >> 6) | 0x00000800;
+                    // fall-through
                 case $c >= 0x00000080:
                     $v .= chr(0x80 | ($c & 0x3F));
                     $c = ($c >> 6) | 0x000000C0;
+                    // fall-through
                 default:
                     $v .= chr($c);
                     break;
@@ -1410,5 +1431,78 @@ class ASN1
             $out .= strrev($v);
         }
         return $out;
+    }
+
+    /**
+     * Extract raw BER from Base64 encoding
+     *
+     * @param string $str
+     * @return string
+     */
+    public static function extractBER($str)
+    {
+        /* X.509 certs are assumed to be base64 encoded but sometimes they'll have additional things in them
+         * above and beyond the ceritificate.
+         * ie. some may have the following preceding the -----BEGIN CERTIFICATE----- line:
+         *
+         * Bag Attributes
+         *     localKeyID: 01 00 00 00
+         * subject=/O=organization/OU=org unit/CN=common name
+         * issuer=/O=organization/CN=common name
+         */
+        if (strlen($str) > ini_get('pcre.backtrack_limit')) {
+            $temp = $str;
+        } else {
+            $temp = preg_replace('#.*?^-+[^-]+-+[\r\n ]*$#ms', '', $str, 1);
+            $temp = preg_replace('#-+END.*[\r\n ]*.*#ms', '', $temp, 1);
+        }
+        // remove new lines
+        $temp = str_replace(["\r", "\n", ' '], '', $temp);
+        // remove the -----BEGIN CERTIFICATE----- and -----END CERTIFICATE----- stuff
+        $temp = preg_replace('#^-+[^-]+-+|-+[^-]+-+$#', '', $temp);
+        $temp = preg_match('#^[a-zA-Z\d/+]*={0,2}$#', $temp) ? Strings::base64_decode($temp) : false;
+        return $temp != false ? $temp : $str;
+    }
+
+    /**
+     * DER-encode the length
+     *
+     * DER supports lengths up to (2**8)**127, however, we'll only support lengths up to (2**8)**4.  See
+     * {@link http://itu.int/ITU-T/studygroups/com17/languages/X.690-0207.pdf#p=13 X.690 paragraph 8.1.3} for more information.
+     *
+     * @param int $length
+     * @return string
+     */
+    public static function encodeLength($length)
+    {
+        if ($length <= 0x7F) {
+            return chr($length);
+        }
+
+        $temp = ltrim(pack('N', $length), chr(0));
+        return pack('Ca*', 0x80 | strlen($temp), $temp);
+    }
+
+    /**
+     * Returns the OID corresponding to a name
+     *
+     * What's returned in the associative array returned by loadX509() (or load*()) is either a name or an OID if
+     * no OID to name mapping is available. The problem with this is that what may be an unmapped OID in one version
+     * of phpseclib may not be unmapped in the next version, so apps that are looking at this OID may not be able
+     * to work from version to version.
+     *
+     * This method will return the OID if a name is passed to it and if no mapping is avialable it'll assume that
+     * what's being passed to it already is an OID and return that instead. A few examples.
+     *
+     * getOID('2.16.840.1.101.3.4.2.1') == '2.16.840.1.101.3.4.2.1'
+     * getOID('id-sha256') == '2.16.840.1.101.3.4.2.1'
+     * getOID('zzz') == 'zzz'
+     *
+     * @param string $name
+     * @return string
+     */
+    public static function getOID($name)
+    {
+        return isset(self::$reverseOIDs[$name]) ? self::$reverseOIDs[$name] : $name;
     }
 }
