@@ -7,7 +7,6 @@ use OCA\OIDCLogin\Service\TokenService;
 use OCP\EventDispatcher\Event;
 use OCP\EventDispatcher\IEventListener;
 use OCP\IConfig;
-use OCP\IRequest;
 use OCP\ISession;
 use OCP\IUserSession;
 use Psr\Log\LoggerInterface;
@@ -15,57 +14,33 @@ use Sabre\DAV\Auth\Backend\AbstractBearer;
 
 class BearerAuthBackend extends AbstractBearer implements IEventListener
 {
-    /** @var string */
-    private $appName;
-
-    /** @var IRequest */
-    private $request;
-
-    /** @var IUserSession */
-    private $userSession;
-
-    /** @var ISession */
-    private $session;
-
-    /** @var IConfig */
-    private $config;
-
-    /** @var string */
-    private $principalPrefix;
-
-    /** @var LoggerInterface */
-    private $logger;
-
-    /** @var LoginService */
-    private $loginService;
-
-    /** @var TokenService */
-    private $tokenService;
+    private string $appName;
+    private IUserSession $userSession;
+    private ISession $session;
+    private IConfig $config;
+    private LoggerInterface $logger;
+    private LoginService $loginService;
+    private string $principalPrefix;
 
     /**
      * @param string $principalPrefix
      */
     public function __construct(
         string $appName,
-        IRequest $request,
         IUserSession $userSession,
         ISession $session,
         IConfig $config,
         LoggerInterface $logger,
         LoginService $loginService,
-        TokenService $tokenService,
         $principalPrefix = 'principals/users/'
     ) {
         $this->appName = $appName;
-        $this->request = $request;
         $this->userSession = $userSession;
         $this->session = $session;
         $this->config = $config;
         $this->logger = $logger;
         $this->loginService = $loginService;
-        $this->tokenService = $tokenService;
         $this->principalPrefix = $principalPrefix;
-        $this->context = ['app' => $appName];
 
         // setup realm
         $defaults = new \OCP\Defaults();
@@ -83,7 +58,7 @@ class BearerAuthBackend extends AbstractBearer implements IEventListener
             try {
                 $this->login($bearerToken);
             } catch (\Exception $e) {
-                $this->logger->debug("WebDAV bearer token validation failed with: {$e->getMessage()}", $this->context);
+                $this->logger->debug("WebDAV bearer token validation failed with: {$e->getMessage()}", ['app' => $this->appName]);
 
                 return false;
             }
@@ -102,15 +77,22 @@ class BearerAuthBackend extends AbstractBearer implements IEventListener
      */
     public function handle(Event $event): void
     {
-        $plugin = $event->getServer()->getPlugin('auth');
-        $webdav_enabled = $this->config->getSystemValue('oidc_login_webdav_enabled', false);
+        if (!$event instanceof \OCA\DAV\Events\SabrePluginAuthInitEvent
+            && !$event instanceof \OCP\SabrePluginEvent) {
+            return;
+        }
 
-        if (null !== $plugin && $webdav_enabled) {
-            $plugin->addBackend($this);
+        $authPlugin = $event->getServer()->getPlugin('auth');
+        if ($authPlugin instanceof \Sabre\DAV\Auth\Plugin) {
+            $webdav_enabled = $this->config->getSystemValue('oidc_login_webdav_enabled', false);
+
+            if ($webdav_enabled) {
+                $authPlugin->addBackend($this);
+            }
         }
     }
 
-    private function setupUserFs($userId)
+    private function setupUserFs(string $userId)
     {
         \OC_Util::setupFS($userId);
         $this->session->close();
@@ -123,9 +105,9 @@ class BearerAuthBackend extends AbstractBearer implements IEventListener
      *
      * @param string $bearerToken an OIDC JWT bearer token
      */
-    private function login(string $bearerToken)
+    public function login(string $bearerToken)
     {
-        $client = $this->tokenService->createOIDCClient();
+        $client = $this->loginService->createOIDCClient();
         if (null === $client) {
             throw new \Exception("Couldn't create OIDC client!");
         }
@@ -134,6 +116,6 @@ class BearerAuthBackend extends AbstractBearer implements IEventListener
 
         $profile = $client->getTokenProfile($bearerToken);
 
-        $this->loginService->login($profile, $this->userSession, $this->request);
+        $this->loginService->login($profile);
     }
 }
