@@ -1,27 +1,25 @@
 <?php
+
 /**
- * Pure-PHP ASN.1 Parser
+ * ASN.1 Raw Element
  *
  * PHP version 5
  *
- * @category  File
- * @package   ASN1
  * @author    Jim Wigginton <terrafrost@php.net>
  * @copyright 2012 Jim Wigginton
  * @license   http://www.opensource.org/licenses/mit-license.html  MIT License
  * @link      http://phpseclib.sourceforge.net
  */
 
-namespace phpseclib\File\ASN1;
+namespace phpseclib3\File\ASN1;
 
 /**
- * ASN.1 Element
+ * ASN.1 Raw Element
  *
- * Bypass normal encoding rules in phpseclib\File\ASN1::encodeDER()
+ * An ASN.1 ANY mapping will return an ASN1\Element object. Use of this object
+ * will also bypass the normal encoding rules in ASN1::encodeDER()
  *
- * @package ASN1
  * @author  Jim Wigginton <terrafrost@php.net>
- * @access  public
  */
 class Element
 {
@@ -29,18 +27,16 @@ class Element
      * Raw element value
      *
      * @var string
-     * @access private
      */
-    var $element;
+    public $element;
 
     /**
      * Constructor
      *
      * @param string $encoded
-     * @return \phpseclib\File\ASN1\Element
-     * @access public
+     * @return \phpseclib3\File\ASN1\Element
      */
-    function __construct($encoded)
+    public function __construct($encoded)
     {
         $this->element = $encoded;
     }
