@@ -7,25 +7,22 @@
  *
  * PHP version 5
  *
- * @category  Net
- * @package   SFTP
  * @author    Jim Wigginton <terrafrost@php.net>
  * @copyright 2013 Jim Wigginton
  * @license   http://www.opensource.org/licenses/mit-license.html  MIT License
  * @link      http://phpseclib.sourceforge.net
  */
 
-namespace phpseclib\Net\SFTP;
+namespace phpseclib3\Net\SFTP;
 
-use phpseclib\Crypt\RSA;
-use phpseclib\Net\SFTP;
+use phpseclib3\Crypt\Common\PrivateKey;
+use phpseclib3\Net\SFTP;
+use phpseclib3\Net\SSH2;
 
 /**
  * SFTP Stream Wrapper
  *
- * @package SFTP
  * @author  Jim Wigginton <terrafrost@php.net>
- * @access  public
  */
 class Stream
 {
@@ -36,90 +33,80 @@ class Stream
      *
      * @var array
      */
-    static $instances;
+    public static $instances;
 
     /**
      * SFTP instance
      *
      * @var object
-     * @access private
      */
-    var $sftp;
+    private $sftp;
 
     /**
      * Path
      *
      * @var string
-     * @access private
      */
-    var $path;
+    private $path;
 
     /**
      * Mode
      *
      * @var string
-     * @access private
      */
-    var $mode;
+    private $mode;
 
     /**
      * Position
      *
      * @var int
-     * @access private
      */
-    var $pos;
+    private $pos;
 
     /**
      * Size
      *
      * @var int
-     * @access private
      */
-    var $size;
+    private $size;
 
     /**
      * Directory entries
      *
      * @var array
-     * @access private
      */
-    var $entries;
+    private $entries;
 
     /**
      * EOF flag
      *
      * @var bool
-     * @access private
      */
-    var $eof;
+    private $eof;
 
     /**
      * Context resource
      *
-     * Technically this needs to be publically accessible so PHP can set it directly
+     * Technically this needs to be publicly accessible so PHP can set it directly
      *
      * @var resource
-     * @access public
      */
-    var $context;
+    public $context;
 
     /**
      * Notification callback function
      *
      * @var callable
-     * @access public
      */
-    var $notification;
+    private $notification;
 
     /**
      * Registers this class as a URL wrapper.
      *
      * @param string $protocol The wrapper name to be registered.
      * @return bool True on success, false otherwise.
-     * @access public
      */
-    static function register($protocol = 'sftp')
+    public static function register($protocol = 'sftp')
     {
         if (in_array($protocol, stream_get_wrappers(), true)) {
             return false;
@@ -130,9 +117,8 @@ class Stream
     /**
      * The Constructor
      *
-     * @access public
      */
-    function __construct()
+    public function __construct()
     {
         if (defined('NET_SFTP_STREAM_LOGGING')) {
             echo "__construct()\r\n";
@@ -149,21 +135,20 @@ class Stream
      *
      * @param string $path
      * @return string
-     * @access private
      */
-    function _parse_path($path)
+    protected function parse_path($path)
     {
         $orig = $path;
-        extract(parse_url($path) + array('port' => 22));
+        extract(parse_url($path) + ['port' => 22]);
         if (isset($query)) {
-            $path.= '?' . $query;
+            $path .= '?' . $query;
         } elseif (preg_match('/(\?|\?#)$/', $orig)) {
-            $path.= '?';
+            $path .= '?';
         }
         if (isset($fragment)) {
-            $path.= '#' . $fragment;
+            $path .= '#' . $fragment;
         } elseif ($orig[strlen($orig) - 1] == '#') {
-            $path.= '#';
+            $path .= '#';
         }
 
         if (!isset($host)) {
@@ -177,13 +162,12 @@ class Stream
             }
         }
 
-        if ($host[0] == '$') {
-            $host = substr($host, 1);
-            global ${$host};
-            if (($$host instanceof SFTP) === false) {
+        if (preg_match('/^{[a-z0-9]+}$/i', $host)) {
+            $host = SSH2::getConnectionByResourceId($host);
+            if ($host === false) {
                 return false;
             }
-            $this->sftp = $$host;
+            $this->sftp = $host;
         } else {
             if (isset($this->context)) {
                 $context = stream_context_get_options($this->context);
@@ -204,7 +188,7 @@ class Stream
             if (isset($context[$scheme]['password'])) {
                 $pass = $context[$scheme]['password'];
             }
-            if (isset($context[$scheme]['privkey']) && $context[$scheme]['privkey'] instanceof RSA) {
+            if (isset($context[$scheme]['privkey']) && $context[$scheme]['privkey'] instanceof PrivateKey) {
                 $pass = $context[$scheme]['privkey'];
             }
 
@@ -212,7 +196,7 @@ class Stream
                 return false;
             }
 
-            // casting $pass to a string is necessary in the event that it's a \phpseclib\Crypt\RSA object
+            // casting $pass to a string is necessary in the event that it's a \phpseclib3\Crypt\RSA object
             if (isset(self::$instances[$host][$port][$user][(string) $pass])) {
                 $this->sftp = self::$instances[$host][$port][$user][(string) $pass];
             } else {
@@ -255,18 +239,17 @@ class Stream
      * @param int $options
      * @param string $opened_path
      * @return bool
-     * @access public
      */
-    function _stream_open($path, $mode, $options, &$opened_path)
+    private function _stream_open($path, $mode, $options, &$opened_path)
     {
-        $path = $this->_parse_path($path);
+        $path = $this->parse_path($path);
 
         if ($path === false) {
             return false;
         }
         $this->path = $path;
 
-        $this->size = $this->sftp->size($path);
+        $this->size = $this->sftp->filesize($path);
         $this->mode = preg_replace('#[bt]$#', '', $mode);
         $this->eof = false;
 
@@ -297,9 +280,8 @@ class Stream
      *
      * @param int $count
      * @return mixed
-     * @access public
      */
-    function _stream_read($count)
+    private function _stream_read($count)
     {
         switch ($this->mode) {
             case 'w':
@@ -329,7 +311,7 @@ class Stream
             $this->eof = true;
             return false;
         }
-        $this->pos+= strlen($result);
+        $this->pos += strlen($result);
 
         return $result;
     }
@@ -338,10 +320,9 @@ class Stream
      * Write to stream
      *
      * @param string $data
-     * @return mixed
-     * @access public
+     * @return int|false
      */
-    function _stream_write($data)
+    private function _stream_write($data)
     {
         switch ($this->mode) {
             case 'r':
@@ -361,7 +342,7 @@ class Stream
         if ($result === false) {
             return false;
         }
-        $this->pos+= strlen($data);
+        $this->pos += strlen($data);
         if ($this->pos > $this->size) {
             $this->size = $this->pos;
         }
@@ -373,9 +354,8 @@ class Stream
      * Retrieve the current position of a stream
      *
      * @return int
-     * @access public
      */
-    function _stream_tell()
+    private function _stream_tell()
     {
         return $this->pos;
     }
@@ -391,9 +371,8 @@ class Stream
      * will return false. do fread($fp, 1) and feof() will then return true.
      *
      * @return bool
-     * @access public
      */
-    function _stream_eof()
+    private function _stream_eof()
     {
         return $this->eof;
     }
@@ -404,21 +383,20 @@ class Stream
      * @param int $offset
      * @param int $whence
      * @return bool
-     * @access public
      */
-    function _stream_seek($offset, $whence)
+    private function _stream_seek($offset, $whence)
     {
         switch ($whence) {
             case SEEK_SET:
-                if ($offset >= $this->size || $offset < 0) {
+                if ($offset < 0) {
                     return false;
                 }
                 break;
             case SEEK_CUR:
-                $offset+= $this->pos;
+                $offset += $this->pos;
                 break;
             case SEEK_END:
-                $offset+= $this->size;
+                $offset += $this->size;
         }
 
         $this->pos = $offset;
@@ -433,11 +411,10 @@ class Stream
      * @param int $option
      * @param mixed $var
      * @return bool
-     * @access public
      */
-    function _stream_metadata($path, $option, $var)
+    private function _stream_metadata($path, $option, $var)
     {
-        $path = $this->_parse_path($path);
+        $path = $this->parse_path($path);
         if ($path === false) {
             return false;
         }
@@ -447,7 +424,9 @@ class Stream
         //     and https://github.com/php/php-src/blob/master/main/php_streams.h#L592
         switch ($option) {
             case 1: // PHP_STREAM_META_TOUCH
-                return $this->sftp->touch($path, $var[0], $var[1]);
+                $time = isset($var[0]) ? $var[0] : null;
+                $atime = isset($var[1]) ? $var[1] : null;
+                return $this->sftp->touch($path, $time, $atime);
             case 2: // PHP_STREAM_OWNER_NAME
             case 3: // PHP_STREAM_GROUP_NAME
                 return false;
@@ -465,9 +444,8 @@ class Stream
      *
      * @param int $cast_as
      * @return resource
-     * @access public
      */
-    function _stream_cast($cast_as)
+    private function _stream_cast($cast_as)
     {
         return $this->sftp->fsock;
     }
@@ -477,9 +455,8 @@ class Stream
      *
      * @param int $operation
      * @return bool
-     * @access public
      */
-    function _stream_lock($operation)
+    private function _stream_lock($operation)
     {
         return false;
     }
@@ -488,15 +465,14 @@ class Stream
      * Renames a file or directory
      *
      * Attempts to rename oldname to newname, moving it between directories if necessary.
-     * If newname exists, it will be overwritten.  This is a departure from what \phpseclib\Net\SFTP
+     * If newname exists, it will be overwritten.  This is a departure from what \phpseclib3\Net\SFTP
      * does.
      *
      * @param string $path_from
      * @param string $path_to
      * @return bool
-     * @access public
      */
-    function _rename($path_from, $path_to)
+    private function _rename($path_from, $path_to)
     {
         $path1 = parse_url($path_from);
         $path2 = parse_url($path_to);
@@ -505,7 +481,7 @@ class Stream
             return false;
         }
 
-        $path_from = $this->_parse_path($path_from);
+        $path_from = $this->parse_path($path_from);
         $path_to = parse_url($path_to);
         if ($path_from === false) {
             return false;
@@ -546,11 +522,10 @@ class Stream
      * @param string $path
      * @param int $options
      * @return bool
-     * @access public
      */
-    function _dir_opendir($path, $options)
+    private function _dir_opendir($path, $options)
     {
-        $path = $this->_parse_path($path);
+        $path = $this->parse_path($path);
         if ($path === false) {
             return false;
         }
@@ -563,9 +538,8 @@ class Stream
      * Read entry from directory handle
      *
      * @return mixed
-     * @access public
      */
-    function _dir_readdir()
+    private function _dir_readdir()
     {
         if (isset($this->entries[$this->pos])) {
             return $this->entries[$this->pos++];
@@ -577,9 +551,8 @@ class Stream
      * Rewind directory handle
      *
      * @return bool
-     * @access public
      */
-    function _dir_rewinddir()
+    private function _dir_rewinddir()
     {
         $this->pos = 0;
         return true;
@@ -589,9 +562,8 @@ class Stream
      * Close directory handle
      *
      * @return bool
-     * @access public
      */
-    function _dir_closedir()
+    private function _dir_closedir()
     {
         return true;
     }
@@ -605,11 +577,10 @@ class Stream
      * @param int $mode
      * @param int $options
      * @return bool
-     * @access public
      */
-    function _mkdir($path, $mode, $options)
+    private function _mkdir($path, $mode, $options)
     {
-        $path = $this->_parse_path($path);
+        $path = $this->parse_path($path);
         if ($path === false) {
             return false;
         }
@@ -626,14 +597,12 @@ class Stream
      * $options. What does 8 correspond to?
      *
      * @param string $path
-     * @param int $mode
      * @param int $options
      * @return bool
-     * @access public
      */
-    function _rmdir($path, $options)
+    private function _rmdir($path, $options)
     {
-        $path = $this->_parse_path($path);
+        $path = $this->parse_path($path);
         if ($path === false) {
             return false;
         }
@@ -644,12 +613,11 @@ class Stream
     /**
      * Flushes the output
      *
-     * See <http://php.net/fflush>. Always returns true because \phpseclib\Net\SFTP doesn't cache stuff before writing
+     * See <http://php.net/fflush>. Always returns true because \phpseclib3\Net\SFTP doesn't cache stuff before writing
      *
      * @return bool
-     * @access public
      */
-    function _stream_flush()
+    private function _stream_flush()
     {
         return true;
     }
@@ -658,9 +626,8 @@ class Stream
      * Retrieve information about a file resource
      *
      * @return mixed
-     * @access public
      */
-    function _stream_stat()
+    private function _stream_stat()
     {
         $results = $this->sftp->stat($this->path);
         if ($results === false) {
@@ -674,11 +641,10 @@ class Stream
      *
      * @param string $path
      * @return bool
-     * @access public
      */
-    function _unlink($path)
+    private function _unlink($path)
     {
-        $path = $this->_parse_path($path);
+        $path = $this->parse_path($path);
         if ($path === false) {
             return false;
         }
@@ -689,18 +655,17 @@ class Stream
     /**
      * Retrieve information about a file
      *
-     * Ignores the STREAM_URL_STAT_QUIET flag because the entirety of \phpseclib\Net\SFTP\Stream is quiet by default
+     * Ignores the STREAM_URL_STAT_QUIET flag because the entirety of \phpseclib3\Net\SFTP\Stream is quiet by default
      * might be worthwhile to reconstruct bits 12-16 (ie. the file type) if mode doesn't have them but we'll
      * cross that bridge when and if it's reached
      *
      * @param string $path
      * @param int $flags
      * @return mixed
-     * @access public
      */
-    function _url_stat($path, $flags)
+    private function _url_stat($path, $flags)
     {
-        $path = $this->_parse_path($path);
+        $path = $this->parse_path($path);
         if ($path === false) {
             return false;
         }
@@ -718,9 +683,8 @@ class Stream
      *
      * @param int $new_size
      * @return bool
-     * @access public
      */
-    function _stream_truncate($new_size)
+    private function _stream_truncate($new_size)
     {
         if (!$this->sftp->truncate($this->path, $new_size)) {
             return false;
@@ -736,15 +700,14 @@ class Stream
      * Change stream options
      *
      * STREAM_OPTION_WRITE_BUFFER isn't supported for the same reason stream_flush isn't.
-     * The other two aren't supported because of limitations in \phpseclib\Net\SFTP.
+     * The other two aren't supported because of limitations in \phpseclib3\Net\SFTP.
      *
      * @param int $option
      * @param int $arg1
      * @param int $arg2
      * @return bool
-     * @access public
      */
-    function _stream_set_option($option, $arg1, $arg2)
+    private function _stream_set_option($option, $arg1, $arg2)
     {
         return false;
     }
@@ -752,9 +715,8 @@ class Stream
     /**
      * Close an resource
      *
-     * @access public
      */
-    function _stream_close()
+    private function _stream_close()
     {
     }
 
@@ -768,12 +730,11 @@ class Stream
      * If NET_SFTP_STREAM_LOGGING is defined all calls will be output on the screen and then (regardless of whether or not
      * NET_SFTP_STREAM_LOGGING is enabled) the parameters will be passed through to the appropriate method.
      *
-     * @param string
-     * @param array
+     * @param string $name
+     * @param array $arguments
      * @return mixed
-     * @access public
      */
-    function __call($name, $arguments)
+    public function __call($name, array $arguments)
     {
         if (defined('NET_SFTP_STREAM_LOGGING')) {
             echo $name . '(';
@@ -790,6 +751,6 @@ class Stream
         if (!method_exists($this, $name)) {
             return false;
         }
-        return call_user_func_array(array($this, $name), $arguments);
+        return $this->$name(...$arguments);
     }
 }
