@@ -1,93 +1,102 @@
 <?php
+
 /**
  * Pure-PHP ssh-agent client.
  *
+ * {@internal See http://api.libssh.org/rfc/PROTOCOL.agent}
+ *
  * PHP version 5
  *
- * @category  System
- * @package   SSH\Agent
  * @author    Jim Wigginton <terrafrost@php.net>
  * @copyright 2009 Jim Wigginton
  * @license   http://www.opensource.org/licenses/mit-license.html  MIT License
  * @link      http://phpseclib.sourceforge.net
- * @internal  See http://api.libssh.org/rfc/PROTOCOL.agent
  */
 
-namespace phpseclib\System\SSH\Agent;
+namespace phpseclib3\System\SSH\Agent;
 
-use phpseclib\System\SSH\Agent;
+use phpseclib3\Common\Functions\Strings;
+use phpseclib3\Crypt\Common\PrivateKey;
+use phpseclib3\Crypt\Common\PublicKey;
+use phpseclib3\Crypt\DSA;
+use phpseclib3\Crypt\EC;
+use phpseclib3\Crypt\RSA;
+use phpseclib3\Exception\UnsupportedAlgorithmException;
+use phpseclib3\System\SSH\Agent;
+use phpseclib3\System\SSH\Common\Traits\ReadBytes;
 
 /**
  * Pure-PHP ssh-agent client identity object
  *
- * Instantiation should only be performed by \phpseclib\System\SSH\Agent class.
- * This could be thought of as implementing an interface that phpseclib\Crypt\RSA
+ * Instantiation should only be performed by \phpseclib3\System\SSH\Agent class.
+ * This could be thought of as implementing an interface that phpseclib3\Crypt\RSA
  * implements. ie. maybe a Net_SSH_Auth_PublicKey interface or something.
  * The methods in this interface would be getPublicKey and sign since those are the
  * methods phpseclib looks for to perform public key authentication.
  *
- * @package SSH\Agent
  * @author  Jim Wigginton <terrafrost@php.net>
- * @access  internal
+ * @internal
  */
-class Identity
+class Identity implements PrivateKey
 {
-    /**@+
-     * Signature Flags
-     *
-     * See https://tools.ietf.org/html/draft-miller-ssh-agent-00#section-5.3
-     *
-     * @access private
-     */
+    use ReadBytes;
+
+    // Signature Flags
+    // See https://tools.ietf.org/html/draft-miller-ssh-agent-00#section-5.3
     const SSH_AGENT_RSA2_256 = 2;
     const SSH_AGENT_RSA2_512 = 4;
-    /**#@-*/
 
     /**
      * Key Object
      *
-     * @var \phpseclib\Crypt\RSA
-     * @access private
+     * @var PublicKey
      * @see self::getPublicKey()
      */
-    var $key;
+    private $key;
 
     /**
      * Key Blob
      *
      * @var string
-     * @access private
      * @see self::sign()
      */
-    var $key_blob;
+    private $key_blob;
 
     /**
      * Socket Resource
      *
      * @var resource
-     * @access private
      * @see self::sign()
      */
-    var $fsock;
+    private $fsock;
 
     /**
      * Signature flags
      *
      * @var int
-     * @access private
      * @see self::sign()
      * @see self::setHash()
      */
-    var $flags = 0;
+    private $flags = 0;
+
+    /**
+     * Curve Aliases
+     *
+     * @var array
+     */
+    private static $curveAliases = [
+        'secp256r1' => 'nistp256',
+        'secp384r1' => 'nistp384',
+        'secp521r1' => 'nistp521',
+        'Ed25519' => 'Ed25519'
+    ];
 
     /**
      * Default Constructor.
      *
      * @param resource $fsock
-     * @return \phpseclib\System\SSH\Agent\Identity
-     * @access private
      */
-    function __construct($fsock)
+    public function __construct($fsock)
     {
         $this->fsock = $fsock;
     }
@@ -95,29 +104,36 @@ class Identity
     /**
      * Set Public Key
      *
-     * Called by \phpseclib\System\SSH\Agent::requestIdentities()
+     * Called by \phpseclib3\System\SSH\Agent::requestIdentities()
      *
-     * @param \phpseclib\Crypt\RSA $key
-     * @access private
+     * @param \phpseclib3\Crypt\Common\PublicKey $key
      */
-    function setPublicKey($key)
+    public function withPublicKey(PublicKey $key)
     {
-        $this->key = $key;
-        $this->key->setPublicKey();
+        if ($key instanceof EC) {
+            if (is_array($key->getCurve()) || !isset(self::$curveAliases[$key->getCurve()])) {
+                throw new UnsupportedAlgorithmException('The only supported curves are nistp256, nistp384, nistp512 and Ed25519');
+            }
+        }
+
+        $new = clone $this;
+        $new->key = $key;
+        return $new;
     }
 
     /**
      * Set Public Key
      *
-     * Called by \phpseclib\System\SSH\Agent::requestIdentities(). The key blob could be extracted from $this->key
+     * Called by \phpseclib3\System\SSH\Agent::requestIdentities(). The key blob could be extracted from $this->key
      * but this saves a small amount of computation.
      *
      * @param string $key_blob
-     * @access private
      */
-    function setPublicKeyBlob($key_blob)
+    public function withPublicKeyBlob($key_blob)
     {
-        $this->key_blob = $key_blob;
+        $new = clone $this;
+        $new->key_blob = $key_blob;
+        return $new;
     }
 
     /**
@@ -125,51 +141,116 @@ class Identity
      *
      * Wrapper for $this->key->getPublicKey()
      *
-     * @param int $format optional
+     * @param string $type optional
      * @return mixed
-     * @access public
      */
-    function getPublicKey($format = null)
+    public function getPublicKey($type = 'PKCS8')
     {
-        return !isset($format) ? $this->key->getPublicKey() : $this->key->getPublicKey($format);
+        return $this->key;
     }
 
     /**
-     * Set Signature Mode
-     *
-     * Doesn't do anything as ssh-agent doesn't let you pick and choose the signature mode. ie.
-     * ssh-agent's only supported mode is \phpseclib\Crypt\RSA::SIGNATURE_PKCS1
-     *
-     * @param int $mode
-     * @access public
-     */
-    function setSignatureMode($mode)
-    {
-    }
-
-    /**
-     * Set Hash
-     *
-     * ssh-agent doesn't support using hashes for RSA other than SHA1
+     * Sets the hash
      *
      * @param string $hash
-     * @access public
      */
-    function setHash($hash)
+    public function withHash($hash)
     {
-        $this->flags = 0;
-        switch ($hash) {
-            case 'sha1':
-                break;
-            case 'sha256':
-                $this->flags = self::SSH_AGENT_RSA2_256;
-                break;
-            case 'sha512':
-                $this->flags = self::SSH_AGENT_RSA2_512;
-                break;
-            default:
-                user_error('The only supported hashes for RSA are sha1, sha256 and sha512');
+        $new = clone $this;
+
+        $hash = strtolower($hash);
+
+        if ($this->key instanceof RSA) {
+            $new->flags = 0;
+            switch ($hash) {
+                case 'sha1':
+                    break;
+                case 'sha256':
+                    $new->flags = self::SSH_AGENT_RSA2_256;
+                    break;
+                case 'sha512':
+                    $new->flags = self::SSH_AGENT_RSA2_512;
+                    break;
+                default:
+                    throw new UnsupportedAlgorithmException('The only supported hashes for RSA are sha1, sha256 and sha512');
+            }
         }
+        if ($this->key instanceof EC) {
+            switch ($this->key->getCurve()) {
+                case 'secp256r1':
+                    $expectedHash = 'sha256';
+                    break;
+                case 'secp384r1':
+                    $expectedHash = 'sha384';
+                    break;
+                //case 'secp521r1':
+                //case 'Ed25519':
+                default:
+                    $expectedHash = 'sha512';
+            }
+            if ($hash != $expectedHash) {
+                throw new UnsupportedAlgorithmException('The only supported hash for ' . self::$curveAliases[$this->key->getCurve()] . ' is ' . $expectedHash);
+            }
+        }
+        if ($this->key instanceof DSA) {
+            if ($hash != 'sha1') {
+                throw new UnsupportedAlgorithmException('The only supported hash for DSA is sha1');
+            }
+        }
+        return $new;
+    }
+
+    /**
+     * Sets the padding
+     *
+     * Only PKCS1 padding is supported
+     *
+     * @param string $padding
+     */
+    public function withPadding($padding)
+    {
+        if (!$this->key instanceof RSA) {
+            throw new UnsupportedAlgorithmException('Only RSA keys support padding');
+        }
+        if ($padding != RSA::SIGNATURE_PKCS1 && $padding != RSA::SIGNATURE_RELAXED_PKCS1) {
+            throw new UnsupportedAlgorithmException('ssh-agent can only create PKCS1 signatures');
+        }
+        return $this;
+    }
+
+    /**
+     * Determines the signature padding mode
+     *
+     * Valid values are: ASN1, SSH2, Raw
+     *
+     * @param string $format
+     */
+    public function withSignatureFormat($format)
+    {
+        if ($this->key instanceof RSA) {
+            throw new UnsupportedAlgorithmException('Only DSA and EC keys support signature format setting');
+        }
+        if ($format != 'SSH2') {
+            throw new UnsupportedAlgorithmException('Only SSH2-formatted signatures are currently supported');
+        }
+
+        return $this;
+    }
+
+    /**
+     * Returns the curve
+     *
+     * Returns a string if it's a named curve, an array if not
+     *
+     * @return string|array
+     */
+    public function getCurve()
+    {
+        if (!$this->key instanceof EC) {
+            throw new UnsupportedAlgorithmException('Only EC keys have curves');
+        }
+
+        return $this->key->getCurve();
     }
 
     /**
@@ -179,63 +260,61 @@ class Identity
      *
      * @param string $message
      * @return string
-     * @access public
+     * @throws \RuntimeException on connection errors
+     * @throws \phpseclib3\Exception\UnsupportedAlgorithmException if the algorithm is unsupported
      */
-    function sign($message)
+    public function sign($message)
     {
         // the last parameter (currently 0) is for flags and ssh-agent only defines one flag (for ssh-dss): SSH_AGENT_OLD_SIGNATURE
-        $packet = pack('CNa*Na*N', Agent::SSH_AGENTC_SIGN_REQUEST, strlen($this->key_blob), $this->key_blob, strlen($message), $message, $this->flags);
-        $packet = pack('Na*', strlen($packet), $packet);
+        $packet = Strings::packSSH2(
+            'CssN',
+            Agent::SSH_AGENTC_SIGN_REQUEST,
+            $this->key_blob,
+            $message,
+            $this->flags
+        );
+        $packet = Strings::packSSH2('s', $packet);
         if (strlen($packet) != fputs($this->fsock, $packet)) {
-            user_error('Connection closed during signing');
-            return false;
+            throw new \RuntimeException('Connection closed during signing');
         }
 
-        $temp = fread($this->fsock, 4);
-        if (strlen($temp) != 4) {
-            user_error('Connection closed during signing');
-            return false;
-        }
-        $length = current(unpack('N', $temp));
-        $type = ord(fread($this->fsock, 1));
+        $length = current(unpack('N', $this->readBytes(4)));
+        $packet = $this->readBytes($length);
+
+        list($type, $signature_blob) = Strings::unpackSSH2('Cs', $packet);
         if ($type != Agent::SSH_AGENT_SIGN_RESPONSE) {
-            user_error('Unable to retrieve signature');
-            return false;
+            throw new \RuntimeException('Unable to retrieve signature');
         }
 
-        $signature_blob = fread($this->fsock, $length - 1);
-        if (strlen($signature_blob) != $length - 1) {
-            user_error('Connection closed during signing');
-            return false;
+        if (!$this->key instanceof RSA) {
+            return $signature_blob;
         }
-        $length = current(unpack('N', $this->_string_shift($signature_blob, 4)));
-        if ($length != strlen($signature_blob)) {
-            user_error('Malformed signature blob');
-        }
-        $length = current(unpack('N', $this->_string_shift($signature_blob, 4)));
-        if ($length > strlen($signature_blob) + 4) {
-            user_error('Malformed signature blob');
-        }
-        $type = $this->_string_shift($signature_blob, $length);
-        $this->_string_shift($signature_blob, 4);
+
+        list($type, $signature_blob) = Strings::unpackSSH2('ss', $signature_blob);
 
         return $signature_blob;
     }
 
     /**
-     * String Shift
+     * Returns the private key
      *
-     * Inspired by array_shift
-     *
-     * @param string $string
-     * @param int $index
+     * @param string $type
+     * @param array $options optional
      * @return string
-     * @access private
      */
-    function _string_shift(&$string, $index = 1)
+    public function toString($type, array $options = [])
     {
-        $substr = substr($string, 0, $index);
-        $string = substr($string, $index);
-        return $substr;
+        throw new \RuntimeException('ssh-agent does not provide a mechanism to get the private key');
+    }
+
+    /**
+     * Sets the password
+     *
+     * @param string|bool $password
+     * @return never
+     */
+    public function withPassword($password = false)
+    {
+        throw new \RuntimeException('ssh-agent does not provide a mechanism to get the private key');
     }
 }
